+++ .github/workflows/CI_M1.yml
@@ -8,7 +8,7 @@ on:
 jobs:
   test:
     name: Julia ${{ matrix.version }} - macOS - ${{ matrix.arch }} - ${{ github.event_name }}
-    runs-on: self-hosted
+    runs-on: [self-hosted, macOS]
     strategy:
       fail-fast: false
       matrix:
