### 2022-12-20 be4514d
[CI] Use Cirrus for tests with Apple M1 instead of CI_M1.yml
