+++ .cirrus.yml
@@ -24,8 +24,15 @@ task:
       env:
         - JULIA_VERSION: 1
   allow_failures: $JULIA_VERSION == 'nightly'
-  install_script:
-    - sh bin/install.sh
+  install_script: |
+    URL="https://raw.githubusercontent.com/ararslan/CirrusCI.jl/master/bin/install.sh"
+    if command -v curl; then
+        sh -c "$(curl ${URL})"
+    elif command -v wget; then
+        sh -c "$(wget ${URL} -q -O-)"
+    elif command -v fetch; then
+        sh -c "$(fetch ${URL} -o -)"
+    fi
   build_script:
     - cirrusjl build
   test_script:
