+++ .cirrus.yml
@@ -7,7 +7,6 @@ task:
         matrix:
           - JULIA_VERSION: 1.6
           - JULIA_VERSION: 1
-          - JULIA_VERSION: nightly
     - name: Linux ARMv8
       arm_container:
         image: ubuntu:latest
@@ -18,16 +17,15 @@ task:
         image: alpine:3.14
       env:
         - JULIA_VERSION: 1
-    - name: macOS M1
+    - name: MacOS M1
       macos_instance:
         image: ghcr.io/cirruslabs/macos-monterey-base:latest
       env:
         - JULIA_VERSION: 1
-  allow_failures: $JULIA_VERSION == 'nightly'
   install_script: |
     URL="https://raw.githubusercontent.com/ararslan/CirrusCI.jl/master/bin/install.sh"
     set -x
-    if command -v apt; then
+    if [ "$(uname -s)" = "Linux" ] && command -v apt; then
         apt update
         apt install -y curl
     fi
