### 2022-12-21 12b4ab6
Use Cirrus to test Krylov.jl with Apple M1
