### 2022-12-23 eb083b8
Switch Metal.jl CI to the juliaecosystem pipeline.
