+++ src/minres.jl
@@ -287,7 +287,7 @@ function minres!(solver :: MinresSolver{T,FC,S}, A, b :: AbstractVector{FC};
 
     # Compute lower bound on forward error.
     err_vec[mod(iter, window) + 1] = ϕ
-    iter ≥ window && (err_lbnd = norm(err_vec))
+    iter ≥ window && (err_lbnd = @knrm2(window, err_vec))
 
     γmax = max(γmax, γ)
     γmin = min(γmin, γ)
