+++ test/test_stats.jl
@@ -4,7 +4,7 @@
   show(io, stats)
   showed = String(take!(io))
   storage_type = typeof(stats)
-  expected = """Simple stats
+  expected = """SimpleStats
   niter: 0
   solved: true
   inconsistent: true
@@ -15,14 +15,15 @@
   @test strip.(split(chomp(showed), "\n")) == strip.(split(chomp(expected), "\n"))
   Krylov.reset!(stats)
   check_reset(stats)
-  @test (VERSION < v"1.5") || (@allocated Krylov.reset!(stats)) == 0
+  nbytes_allocated = @allocated Krylov.reset!(stats)
+  @test nbytes_allocated == 0
 
   stats = Krylov.LsmrStats(0, true, true, Float64[1.0], Float64[2.0], Float64(3.0), Float64(4.0), Float64(5.0), Float64(6.0), Float64(7.0), "t")
   io = IOBuffer()
   show(io, stats)
   showed = String(take!(io))
   storage_type = typeof(stats)
-  expected = """Lsmr stats
+  expected = """LsmrStats
   niter: 0
   solved: true
   inconsistent: true
@@ -37,14 +38,15 @@
   @test strip.(split(chomp(showed), "\n")) == strip.(split(chomp(expected), "\n"))
   Krylov.reset!(stats)
   check_reset(stats)
-  @test (VERSION < v"1.5") || (@allocated Krylov.reset!(stats)) == 0
+  nbytes_allocated = @allocated Krylov.reset!(stats)
+  @test nbytes_allocated == 0
 
   stats = Krylov.LanczosStats(0, true, Float64[3.0], true, NaN, NaN, "t")
   io = IOBuffer()
   show(io, stats)
   showed = String(take!(io))
   storage_type = typeof(stats)
-  expected = """Lanczos stats
+  expected = """LanczosStats
   niter: 0
   solved: true
   residuals: [ 3.0e+00 ]
@@ -55,14 +57,15 @@
   @test strip.(split(chomp(showed), "\n")) == strip.(split(chomp(expected), "\n"))
   Krylov.reset!(stats)
   check_reset(stats)
-  @test (VERSION < v"1.5") || (@allocated Krylov.reset!(stats)) == 0
+  nbytes_allocated = @allocated Krylov.reset!(stats)
+  @test nbytes_allocated == 0
 
   stats = Krylov.LanczosShiftStats(0, true, [Float64[0.9, 0.5], Float64[0.6, 0.4, 0.1]], BitVector([false, true]), NaN, NaN, "t")
   io = IOBuffer()
   show(io, stats)
   showed = String(take!(io))
   storage_type = typeof(stats)
-  expected = """LanczosShift stats
+  expected = """LanczosShiftStats
   niter: 0
   solved: true
   residuals: [[0.9, 0.5], [0.6, 0.4, 0.1]]
@@ -70,16 +73,17 @@
   ‖A‖F: NaN
   κ₂(A): NaN
   status: t"""
-  @test (VERSION < v"1.5") || strip.(split(chomp(showed), "\n")) == strip.(split(chomp(expected), "\n"))
+  @test strip.(split(chomp(showed), "\n")) == strip.(split(chomp(expected), "\n"))
   Krylov.reset!(stats)
-  @test (VERSION < v"1.5") || (@allocated Krylov.reset!(stats)) == 0
+  nbytes_allocated = @allocated Krylov.reset!(stats)
+  @test nbytes_allocated == 0
 
   stats = Krylov.SymmlqStats(0, true, Float64[4.0], Union{Float64,Missing}[5.0, missing], Float64[6.0], Union{Float64,Missing}[7.0, missing], NaN, NaN, "t")
   io = IOBuffer()
   show(io, stats)
[…]
