+++ src/krylov_stats.jl
@@ -24,6 +24,12 @@ mutable struct SimpleStats{T} <: KrylovStats{T}
   status       :: String
 end
 
+function reset!(stats :: SimpleStats)
+  empty!(stats.residuals)
+  empty!(stats.Aresiduals)
+  empty!(stats.Acond)
+end
+
 """
 Type for statistics returned by LSMR. The attributes are:
 - niter
@@ -50,6 +56,11 @@ mutable struct LsmrStats{T} <: KrylovStats{T}
   status       :: String
 end
 
+function reset!(stats :: LsmrStats)
+  empty!(stats.residuals)
+  empty!(stats.Aresiduals)
+end
+
 """
 Type for statistics returned by CG-LANCZOS, the attributes are:
 - niter
@@ -68,6 +79,10 @@ mutable struct LanczosStats{T} <: KrylovStats{T}
   Anorm      :: T
   Acond      :: T
   status     :: String
+end
+
+function reset!(stats :: LanczosStats)
+  empty!(stats.residuals)
 end
 
 """
@@ -120,6 +135,13 @@ mutable struct SymmlqStats{T} <: KrylovStats{T}
   status      :: String
 end
 
+function reset!(stats :: SymmlqStats)
+  empty!(stats.residuals)
+  empty!(stats.residualscg)
+  empty!(stats.errors)
+  empty!(stats.errorscg)
+end
+
 """
 Type for statistics returned by adjoint systems solvers BiLQR and TriLQR, the attributes are:
 - niter
@@ -136,6 +158,11 @@ mutable struct AdjointStats{T} <: KrylovStats{T}
   residuals_primal :: Vector{T}
   residuals_dual   :: Vector{T}
   status           :: String
+end
+
+function reset!(stats :: AdjointStats)
+  empty!(stats.residuals_primal)
+  empty!(stats.residuals_dual)
 end
 
 """
@@ -156,6 +183,12 @@ mutable struct LNLQStats{T} <: KrylovStats{T}
   error_bnd_x    :: Vector{T}
   error_bnd_y    :: Vector{T}
   status         :: String
+end
+
+function reset!(stats :: LNLQStats)
+  empty!(stats.residuals)
+  empty!(stats.error_bnd_x)
+  empty!(stats.error_bnd_y)
 end
 
 """
@@ -184,6 +217,14 @@ mutable struct LSLQStats{T} <: KrylovStats{T}
   status         :: String
 end
 
+function reset!(stats :: LSLQStats)
+  empty!(stats.residuals)
+  empty!(stats.Aresiduals)
+  empty!(stats.err_lbnds)
+  empty!(stats.err_ubnds_lq)
+  empty!(stats.err_ubnds_cg)
+end
+
 import Base.show
 
 special_fields = Dict(
@@ -195,45 +236,24 @@ special_fields = Dict(
   :err_ubnds_cg => "error bound CG",
 )
 
-for f in ["Simple", "Lsmr", "Adjoint", "LNLQ", "LSLQ", "Lanczos", "Symmlq"]
-  T = Meta.parse("Krylov." * f * "Stats{S}")
-
-  @eval function empty_field!(stats :: $T, i, ::Type{Vector{Si}}) where {S, Si}
-    statfield = getfield(stats, i)
-    empty!(statfield)
-  end
-  @eval empty_field!(stats :: $T, i, type) where S = stats
-
-  @eval function reset!(stats :: $T) where S
-    nfield = length($T.types)
-    for i = 1 : nfield
-      type  = fieldtype($T, i)
-      empty_field!(stats, i, type)
+function show(io :: IO, stats :: KrylovStats)
+  kst = typeof(stats)
+  s = string(kst.name.name) * "\n"
+  nfield = fieldcount(kst)
+  for i = 1 : nfield
[…]
