### 2023-01-03 #4970 @jedbrown
Thanks @amontoison for your submission and sorry about the delay. I've surveyed your submission and it looks great. I would ask that you discuss related work:
.
> * **A statement of need:** Does the paper have a section titled ‘Statement of need’ that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?
> * **State of the field:** Do the authors describe how this software compares to other commonly-used packages?
>
> https://joss.readthedocs.io/en/latest/review_checklist.html#software-paper

In pure numbers, I think PETSc has more Krylov methods, but probably not if you consider flexible and pipelined/communication-amortizing ones to all be the same. I think this statement needs some discussion/nuance.
