### 2023-01-03 #4970 @jedbrown
@leios @mvaldes :wave: Would you be available to review this submission to JOSS?
