### 2023-01-06 #4970 @amontoison
Thanks @jedbrown for your answer.
For the section `Statement of need`, we preferred to title it `Features and Functionalities` for information but we can rename it if it's mandatory for JOSS.

The two other main Julia packages that contains Krylov methods are [IterativeSolvers.jl](https://github.com/JuliaLinearAlgebra/IterativeSolvers.jl) and [KrylovKit.jl](https://github.com/Jutho/KrylovKit.jl)  but they are not only focused on linear solvers are have a very limited subset of Krylov methods like cg, lsmr, gmres.
KrylovKit.jl doesn't support preconditioners and IterativeSolvers.jl is only maintained, no new feature since 2/3 years.
Add benchmark with them is not very relevant from my point of view.
However comparisons / benchmarks between Krylov.jl and PETSc could be a nice addition to the paper.

For the number of methods, I didn't know that PETCs have many flexible and pipelined/communication-amortizing variants. From my point of view, we have different Krylov methods if they generate different iterates. I checked with the number of Krylov methods available in MATLAB when I wrote that Krylov.jl has the largest collection of Krylov processes / methods. I will clarify and nuance my statement.
