### 2023-01-06 #4970 @amontoison
Related topic for benchmarks between Krylov.jl and PETCs:
https://discourse.julialang.org/t/creating-and-solving-block-sparse-matrices-with-petsc/90831
The main issue is to do fair comparisons with the same preconditioners and options.
