### 2023-01-22 962811e
Add new tests for GPUs
