+++ test/gpu/intel.jl
@@ -2,12 +2,6 @@ using oneAPI
 
 include("gpu.jl")
 
-import Krylov.kdot
-# https://github.com/JuliaGPU/GPUArrays.jl/pull/427
-function kdot(n :: Integer, x :: oneVector{T}, dx :: Integer, y :: oneVector{T}, dy :: Integer) where T <: Krylov.FloatOrComplex
-  return mapreduce(dot, +, x, y)
-end
-
 @testset "Intel -- oneAPI.jl" begin
 
   @test oneAPI.functional()
@@ -72,9 +66,9 @@ end
       Krylov.@kswap(x, y)
     end
 
-    # @testset "kref! -- $FC" begin
-    #   Krylov.@kref!(n, x, y, c, s)
-    # end
+    @testset "kref! -- $FC" begin
+      Krylov.@kref!(n, x, y, c, s)
+    end
 
     @testset "conversion -- $FC" begin
       test_conversion(S, M)
@@ -86,17 +80,25 @@ end
 
     @testset "GMRES -- $FC" begin
       A, b = nonsymmetric_indefinite(FC=FC)
-      A = oneMatrix{FC}(A)
-      b = oneVector{FC}(b)
+      A = M(A)
+      b = S(b)
       x, stats = gmres(A, b)
       @test norm(b - A * x) ≤ atol + rtol * norm(b)
     end
 
     @testset "CG -- $FC" begin
       A, b = symmetric_definite(FC=FC)
-      A = oneMatrix{FC}(A)
-      b = oneVector{FC}(b)
+      A = M(A)
+      b = S(b)
       x, stats = cg(A, b)
+      @test norm(b - A * x) ≤ atol + rtol * norm(b)
+    end
+
+    @testset "MINRES-QLP -- $FC" begin
+      A, b = symmetric_indefinite(FC=FC)
+      A = M(A)
+      b = S(b)
+      x, stats = minres_qlp(A, b)
       @test norm(b - A * x) ≤ atol + rtol * norm(b)
     end
 
