### 2023-01-22 529dc64
Add new tests for GPUs
