### 2023-01-22 9e4b6da
Use sparse_laplacian instead of symmetric_indefinite for GPU tests
