### 2023-01-22 03ee0a8
[buildkite] Use Julia 1.9 for AMD GPUs
