### 2023-01-22 0bb8b92
Use symmetric_definite instead of sparse_laplacian for GPU tests
