### 2023-01-22 baa45b3
Use symmetric_definite instead of sparse_laplacian for GPU tests
