### 2023-01-22 67dbbee
Debug minres_qlp with verbose mode
