### 2023-01-22 e76d64e
Relax the tolerance of the stopping condition based on the backward error
