### 2023-01-22 1fac754
Use master branch of AMDGPU.jl
