+++ .buildkite/pipeline.yml
@@ -29,7 +29,8 @@ steps:
     command: |
       julia --color=yes --project -e '
       using Pkg
-      Pkg.add("AMDGPU")
+      # Pkg.add("AMDGPU")
+      Pkg.add(url="https://github.com/JuliaGPU/AMDGPU.jl", rev="master")
       Pkg.instantiate()
       include("test/gpu/amd.jl")'
     timeout_in_minutes: 30
