+++ paper/paper.bib
@@ -78,3 +78,19 @@
   doi       = {10.1137/1.9781611970937},
   publisher = {SIAM}
 }
+
+@techreport{petsc,
+  author      = {Satish Balay and Shrirang Abhyankar and Mark~F. Adams and Steven Benson and Jed Brown and Peter Brune and Kris Buschelman and Emil Constantinescu and Lisandro Dalcin and Alp Dener and Victor Eijkhout and Jacob Faibussowitsch and William~D. Gropp and V\'{a}clav Hapla and Tobin Isaac and Pierre Jolivet and Dmitry Karpeev and Dinesh Kaushik and Matthew~G. Knepley and Fande Kong and Scott Kruger and Dave~A. May and Lois Curfman McInnes and Richard Tran Mills and Lawrence Mitchell and Todd Munson and Jose~E. Roman and Karl Rupp and Patrick Sanan and Jason Sarich and Barry~F. Smith and Stefano Zampini and Hong Zhang and Hong Zhang and Junchao Zhang},
+  title       = {{PETSc/TAO} Users Manual},
+  institution = {Argonne National Laboratory},
+  number      = {ANL-21/39 - Revision 3.18},
+  year        = {2022},
+}
+
+@book{MATLAB,
+year = {2022},
+author = {MATLAB},
+title = {version 9.13.0 (R2022b)},
+publisher = {The MathWorks Inc.},
+address = {Natick, Massachusetts}
+}
