### 2023-01-29 #4970 @amontoison
@jedbrown
I renamed the section `Features and Functionalities` into `Statement of need`.
I also added clarifications about the statement that Krylov.jl has the biggest collection of Krylov methods.
I checked the documentation of [PETSc](https://petsc.org/release/overview/linear_solve_table/#krylov-methods) and [MATLAB](https://www.mathworks.com/help/matlab/sparse-matrices.html?s_tid=srchbrcm) (section __Iterative Methods and Preconditioners__) and they have 18 and 11 Krylov methods, respectively.

I still need to add benchmarks between PETCs and Krylov.jl, I will try to find time to do it this week.
