+++ src/lsmr.jl
@@ -35,7 +35,7 @@ export lsmr, lsmr!
                       verbose::Int=0, history::Bool=false,
                       callback=solver->false, iostream::IO=kstdout)
 
-`T` is a `Real` such as `Float32`, `Float64` or `BigFloat`.
+`T` is an `AbstractFloat` such as `Float32`, `Float64` or `BigFloat`.
 `FC` is `T` or `Complex{T}`.
 
 Solve the regularized linear least-squares problem
@@ -121,7 +121,7 @@ In this case, `N` can still be specified and indicates the weighted norm in whic
 """
 function lsmr end
 
-function lsmr(A, b :: AbstractVector{FC}; window :: Int=5, kwargs...) where FC <: RealOrComplex
+function lsmr(A, b :: AbstractVector{FC}; window :: Int=5, kwargs...) where FC <: FloatOrComplex
   solver = LsmrSolver(A, b, window=window)
   lsmr!(solver, A, b; kwargs...)
   return (solver.x, solver.stats)
