+++ src/dqgmres.jl
@@ -18,7 +18,7 @@ export dqgmres, dqgmres!
                          verbose::Int=0, history::Bool=false,
                          callback=solver->false, iostream::IO=kstdout)
 
-`T` is a `Real` such as `Float32`, `Float64` or `BigFloat`.
+`T` is an `AbstractFloat` such as `Float32`, `Float64` or `BigFloat`.
 `FC` is `T` or `Complex{T}`.
 
     (x, stats) = dqgmres(A, b, x0::AbstractVector; kwargs...)
@@ -70,13 +70,13 @@ Otherwise, DQGMRES interpolates between MINRES and GMRES and is similar to MINRE
 """
 function dqgmres end
 
-function dqgmres(A, b :: AbstractVector{FC}, x0 :: AbstractVector; memory :: Int=20, kwargs...) where FC <: RealOrComplex
+function dqgmres(A, b :: AbstractVector{FC}, x0 :: AbstractVector; memory :: Int=20, kwargs...) where FC <: FloatOrComplex
   solver = DqgmresSolver(A, b, memory)
   dqgmres!(solver, A, b, x0; kwargs...)
   return (solver.x, solver.stats)
 end
 
-function dqgmres(A, b :: AbstractVector{FC}; memory :: Int=20, kwargs...) where FC <: RealOrComplex
+function dqgmres(A, b :: AbstractVector{FC}; memory :: Int=20, kwargs...) where FC <: FloatOrComplex
   solver = DqgmresSolver(A, b, memory)
   dqgmres!(solver, A, b; kwargs...)
   return (solver.x, solver.stats)
@@ -95,7 +95,7 @@ See [`DqgmresSolver`](@ref) for more details about the `solver`.
 """
 function dqgmres! end
 
-function dqgmres!(solver :: DqgmresSolver{T,FC,S}, A, b :: AbstractVector{FC}, x0 :: AbstractVector; kwargs...) where {T <: Real, FC <: RealOrComplex{T}, S <: DenseVector{FC}}
+function dqgmres!(solver :: DqgmresSolver{T,FC,S}, A, b :: AbstractVector{FC}, x0 :: AbstractVector; kwargs...) where {T <: AbstractFloat, FC <: FloatOrComplex{T}, S <: DenseVector{FC}}
   warm_start!(solver, x0)
   dqgmres!(solver, A, b; kwargs...)
   return solver
