+++ src/krylov_utils.jl
@@ -164,7 +164,7 @@ Display an array in the form
 
 with (ndisp - 1)/2 elements on each side.
 """
-function vec2str(x :: AbstractVector{T}; ndisp :: Int=7) where T <: Union{Real, Missing}
+function vec2str(x :: AbstractVector{T}; ndisp :: Int=7) where T <: Union{AbstractFloat, Missing}
   n = length(x)
   if n ≤ ndisp
     ndisp = n
