+++ src/krylov_solvers.jl
@@ -73,11 +73,6 @@ mutable struct MinresSolver{T,FC,S} <: KrylovSolver{T,FC,S}
   err_vec    :: Vector{T}
   warm_start :: Bool
   stats      :: SimpleStats{T}
-
-  function MinresSolver{T,FC,S}(Δx, x, r1, r2, w1, w2, y, v, err_vec, warm_start, stats) where {T,FC,S}
-    T <: Integer && error("Krylov methods cannot solve linear systems over the integers")
-    return new(Δx, x, r1, r2, w1, w2, y, v, err_vec, warm_start, stats)
-  end
 end
 
 function MinresSolver(m, n, S; window :: Int=5)
@@ -92,7 +87,6 @@ function MinresSolver(m, n, S; window :: Int=5)
   y  = S(undef, n)
   v  = S(undef, 0)
   err_vec = zeros(T, window)
-  warm_start = false
   stats = SimpleStats(0, false, false, T[], T[], T[], "unknown")
   solver = MinresSolver{T,FC,S}(m, n, Δx, x, r1, r2, w1, w2, y, v, err_vec, false, stats)
   return solver
@@ -125,11 +119,6 @@ mutable struct CgSolver{T,FC,S} <: KrylovSolver{T,FC,S}
   z          :: S
   warm_start :: Bool
   stats      :: SimpleStats{T}
-
-  function CgSolver{T,FC,S}(Δx, x, r, p, Ap, z, warm_start, stats) where {T,FC,S}
-    T <: Integer && error("Krylov methods cannot solve linear systems over the integers")
-    return new(Δx, x, r, p, Ap, z, warm_start, stats)
-  end
 end
 
 function CgSolver(m, n, S)
@@ -141,7 +130,6 @@ function CgSolver(m, n, S)
   p  = S(undef, n)
   Ap = S(undef, n)
   z  = S(undef, 0)
-  warm_start = false
   stats = SimpleStats(0, false, false, T[], T[], T[], "unknown")
   solver = CgSolver{T,FC,S}(m, n, Δx, x, r, p, Ap, z, false, stats)
   return solver
@@ -175,11 +163,6 @@ mutable struct CrSolver{T,FC,S} <: KrylovSolver{T,FC,S}
   Mq         :: S
   warm_start :: Bool
   stats      :: SimpleStats{T}
-
-  function CrSolver{T,FC,S}(Δx, x, r, p, q, Ar, Mq, warm_start, stats) where {T,FC,S}
-    T <: Integer && error("Krylov methods cannot solve linear systems over the integers")
-    return new(Δx, x, r, p, q, Ar, Mq, warm_start, stats)
-  end
 end
 
 function CrSolver(m, n, S)
@@ -192,7 +175,6 @@ function CrSolver(m, n, S)
   q  = S(undef, n)
   Ar = S(undef, n)
   Mq = S(undef, 0)
-  warm_start = false
   stats = SimpleStats(0, false, false, T[], T[], T[], "unknown")
   solver = CrSolver{T,FC,S}(m, n, Δx, x, r, p, q, Ar, Mq, false, stats)
   return solver
@@ -229,11 +211,6 @@ mutable struct SymmlqSolver{T,FC,S} <: KrylovSolver{T,FC,S}
   sprod      :: Vector{T}
   warm_start :: Bool
[…]
