### 2023-02-02 489fcfb
Add a type for some arguments of KrylovSolvers
