+++ src/krylov_solvers.jl
@@ -54,8 +54,8 @@ Type for storing the vectors required by the in-place version of MINRES.
 
 The outer constructors
 
-    solver = MinresSolver(m, n, S; window :: Int=5)
-    solver = MinresSolver(A, b; window :: Int=5)
+    solver = MinresSolver(m, n, S; window :: Integer=5)
+    solver = MinresSolver(A, b; window :: Integer=5)
 
 may be used in order to create these vectors.
 """
@@ -75,7 +75,7 @@ mutable struct MinresSolver{T,FC,S} <: KrylovSolver{T,FC,S}
   stats      :: SimpleStats{T}
 end
 
-function MinresSolver(m, n, S; window :: Int=5)
+function MinresSolver(m :: Integer, n :: Integer, S :: DataType; window :: Integer=5)
   FC = eltype(S)
   T  = real(FC)
   Δx = S(undef, 0)
@@ -92,7 +92,7 @@ function MinresSolver(m, n, S; window :: Int=5)
   return solver
 end
 
-function MinresSolver(A, b; window :: Int=5)
+function MinresSolver(A, b :: AbstractVector; window :: Integer=5)
   m, n = size(A)
   S = ktypeof(b)
   MinresSolver(m, n, S, window=window)
@@ -121,7 +121,7 @@ mutable struct CgSolver{T,FC,S} <: KrylovSolver{T,FC,S}
   stats      :: SimpleStats{T}
 end
 
-function CgSolver(m, n, S)
+function CgSolver(m :: Integer, n :: Integer, S :: DataType)
   FC = eltype(S)
   T  = real(FC)
   Δx = S(undef, 0)
@@ -135,7 +135,7 @@ function CgSolver(m, n, S)
   return solver
 end
 
-function CgSolver(A, b)
+function CgSolver(A, b :: AbstractVector)
   m, n = size(A)
   S = ktypeof(b)
   CgSolver(m, n, S)
@@ -165,7 +165,7 @@ mutable struct CrSolver{T,FC,S} <: KrylovSolver{T,FC,S}
   stats      :: SimpleStats{T}
 end
 
-function CrSolver(m, n, S)
+function CrSolver(m :: Integer, n :: Integer, S :: DataType)
   FC = eltype(S)
   T  = real(FC)
   Δx = S(undef, 0)
@@ -180,7 +180,7 @@ function CrSolver(m, n, S)
   return solver
 end
 
-function CrSolver(A, b)
+function CrSolver(A, b :: AbstractVector)
   m, n = size(A)
   S = ktypeof(b)
   CrSolver(m, n, S)
@@ -213,7 +213,7 @@ mutable struct SymmlqSolver{T,FC,S} <: KrylovSolver{T,FC,S}
   stats      :: SymmlqStats{T}
 end
 
-function SymmlqSolver(m, n, S; window :: Int=5)
+function SymmlqSolver(m :: Integer, n :: Integer, S :: DataType; window :: Integer=5)
   FC      = eltype(S)
   T       = real(FC)
   Δx      = S(undef, 0)
@@ -231,7 +231,7 @@ function SymmlqSolver(m, n, S; window :: Int=5)
   return solver
 end
 
-function SymmlqSolver(A, b; window :: Int=5)
+function SymmlqSolver(A, b :: AbstractVector; window :: Integer=5)
   m, n = size(A)
   S = ktypeof(b)
   SymmlqSolver(m, n, S, window=window)
@@ -261,7 +261,7 @@ mutable struct CgLanczosSolver{T,FC,S} <: KrylovSolver{T,FC,S}
[…]
