### 2023-02-05 21bdd3e
Add more details about the number of Krylov methods
