### 2023-02-09 3607539
Allow more generic storage type S
