### 2023-02-09 c214873
Allow more generic storage type S
