+++ src/krylov_solvers.jl
@@ -75,17 +75,18 @@ mutable struct MinresSolver{T,FC,S} <: KrylovSolver{T,FC,S}
   stats      :: SimpleStats{T}
 end
 
-function MinresSolver(m, n, S; window :: Int=5)
+function MinresSolver(m, n, ::Type{S}; ixm=m, ixn=n, window :: Int=5) where S <: AbstractVector
+  dense = S <: DenseVector
   FC = eltype(S)
   T  = real(FC)
-  Δx = S(undef, 0)
-  x  = S(undef, n)
-  r1 = S(undef, n)
-  r2 = S(undef, n)
-  w1 = S(undef, n)
-  w2 = S(undef, n)
-  y  = S(undef, n)
-  v  = S(undef, 0)
+  Δx = dense ? similar(S, 0) : similar(S, ixn)
+  x  = similar(S, ixn)
+  r1 = similar(S, ixn)
+  r2 = similar(S, ixn)
+  w1 = similar(S, ixn)
+  w2 = similar(S, ixn)
+  y  = similar(S, ixn)
+  v  = dense ? similar(S, 0) : similar(S, ixn)
   err_vec = zeros(T, window)
   stats = SimpleStats(0, false, false, T[], T[], T[], "unknown")
   solver = MinresSolver{T,FC,S}(m, n, Δx, x, r1, r2, w1, w2, y, v, err_vec, false, stats)
@@ -95,7 +96,12 @@ end
 function MinresSolver(A, b; window :: Int=5)
   m, n = size(A)
   S = ktypeof(b)
-  MinresSolver(m, n, S, window=window)
+  if S <: DenseVector
+    ixm, ixn = m, n
+  else
+    ixm, ixn = axes(A)
+  end
+  MinresSolver(m, n, S, ixm=ixm, ixn=ixn, window=window)
 end
 
 """
@@ -121,15 +127,16 @@ mutable struct CgSolver{T,FC,S} <: KrylovSolver{T,FC,S}
   stats      :: SimpleStats{T}
 end
 
-function CgSolver(m, n, S)
+function CgSolver(m, n, ::Type{S}; ixm=m, ixn=n) where S <: AbstractVector
+  dense = S <: DenseVector
   FC = eltype(S)
   T  = real(FC)
-  Δx = S(undef, 0)
-  x  = S(undef, n)
-  r  = S(undef, n)
-  p  = S(undef, n)
-  Ap = S(undef, n)
-  z  = S(undef, 0)
+  Δx = dense ? similar(S, 0) : similar(S, ixn)
+  x  = similar(S, ixn)
+  r  = similar(S, ixn)
+  p  = similar(S, ixn)
+  Ap = similar(S, ixn)
+  z  = dense ? similar(S, 0) : similar(S, ixn)
   stats = SimpleStats(0, false, false, T[], T[], T[], "unknown")
   solver = CgSolver{T,FC,S}(m, n, Δx, x, r, p, Ap, z, false, stats)
   return solver
@@ -138,7 +145,12 @@ end
 function CgSolver(A, b)
   m, n = size(A)
   S = ktypeof(b)
-  CgSolver(m, n, S)
+  if S <: DenseVector
+    ixm, ixn = m, n
+  else
+    ixm, ixn = axes(A)
+  end
+  CgSolver(m, n, S, ixm=ixm, ixn=ixn)
 end
 
 """
@@ -165,16 +177,17 @@ mutable struct CrSolver{T,FC,S} <: KrylovSolver{T,FC,S}
   stats      :: SimpleStats{T}
 end
[…]
