+++ src/krylov_solvers.jl
@@ -96,11 +96,7 @@ end
 function MinresSolver(A, b; window :: Int=5)
   m, n = size(A)
   S = ktypeof(b)
-  if S <: DenseVector
-    ixm, ixn = m, n
-  else
-    ixm, ixn = axes(A)
-  end
+  ixm, ixn = kaxes(S, A)
   MinresSolver(m, n, S, ixm=ixm, ixn=ixn, window=window)
 end
 
@@ -145,11 +141,7 @@ end
 function CgSolver(A, b)
   m, n = size(A)
   S = ktypeof(b)
-  if S <: DenseVector
-    ixm, ixn = m, n
-  else
-    ixm, ixn = axes(A)
-  end
+  ixm, ixn = kaxes(S, A)
   CgSolver(m, n, S, ixm=ixm, ixn=ixn)
 end
 
@@ -196,11 +188,7 @@ end
 function CrSolver(A, b)
   m, n = size(A)
   S = ktypeof(b)
-  if S <: DenseVector
-    ixm, ixn = m, n
-  else
-    ixm, ixn = axes(A)
-  end
+  ixm, ixn = kaxes(S, A)
   CrSolver(m, n, S, ixm=ixm, ixn=ixn)
 end
 
@@ -253,11 +241,7 @@ end
 function SymmlqSolver(A, b; window :: Int=5)
   m, n = size(A)
   S = ktypeof(b)
-  if S <: DenseVector
-    ixm, ixn = m, n
-  else
-    ixm, ixn = axes(A)
-  end
+  ixm, ixn = kaxes(S, A)
   SymmlqSolver(m, n, S, ixm=ixm, ixn=ixn, window=window)
 end
 
@@ -304,11 +288,7 @@ end
 function CgLanczosSolver(A, b)
   m, n = size(A)
   S = ktypeof(b)
-  if S <: DenseVector
-    ixm, ixn = m, n
-  else
-    ixm, ixn = axes(A)
-  end
+  ixm, ixn = kaxes(S, A)
   CgLanczosSolver(m, n, S, ixm=ixm, ixn=ixn)
 end
 
@@ -368,11 +348,7 @@ end
 function CgLanczosShiftSolver(A, b, nshifts)
   m, n = size(A)
   S = ktypeof(b)
-  if S <: DenseVector
-    ixm, ixn = m, n
-  else
-    ixm, ixn = axes(A)
-  end
+  ixm, ixn = kaxes(S, A)
   CgLanczosShiftSolver(m, n, nshifts, S, ixm=ixm, ixn=ixn)
 end
 
@@ -421,11 +397,7 @@ end
 function MinresQlpSolver(A, b)
   m, n = size(A)
   S = ktypeof(b)
-  if S <: DenseVector
-    ixm, ixn = m, n
-  else
-    ixm, ixn = axes(A)
-  end
+  ixm, ixn = kaxes(S, A)
   MinresQlpSolver(m, n, S, ixm=ixm, ixn=ixn)
 end
 
@@ -480,11 +452,7 @@ end
 function DqgmresSolver(A, b, memory = 20)
   m, n = size(A)
   S = ktypeof(b)
-  if S <: DenseVector
-    ixm, ixn = m, n
-  else
-    ixm, ixn = axes(A)
-  end
+  ixm, ixn = kaxes(S, A)
   DqgmresSolver(m, n, memory, S, ixm=ixm, ixn=ixn)
 end
 
@@ -537,11 +505,7 @@ end
 function DiomSolver(A, b, memory = 20)
   m, n = size(A)
   S = ktypeof(b)
-  if S <: DenseVector
[…]
