+++ docs/src/reference.md
@@ -11,5 +11,7 @@ Krylov.FloatOrComplex
 Krylov.niterations
 Krylov.Aprod
 Krylov.Atprod
+Krylov.kaxe
+Krylov.kaxes
 Base.show
 ```
