### 2023-02-11 c8c0836
Check that the KrylovSolvers are compatible with the dimension of the linear problems
