### 2023-02-13 49bdbec
Update solver tests for Julia 1.6
