+++ .buildkite/pipeline.yml
@@ -15,25 +15,24 @@ steps:
       include("test/gpu/nvidia.jl")'
     timeout_in_minutes: 30
 
-  - label: "AMD GPUs -- AMDGPU.jl"
-    plugins:
-      - JuliaCI/julia#v1:
-          version: 1.9-nightly
-    agents:
-      queue: "juliagpu"
-      rocm: "*"
-      rocmgpu: "*"
-    env:
-      JULIA_AMDGPU_CORE_MUST_LOAD: "1"
-      JULIA_AMDGPU_HIP_MUST_LOAD: "1"
-    command: |
-      julia --color=yes --project -e '
-      using Pkg
-      # Pkg.add("AMDGPU")
-      Pkg.add(url="https://github.com/JuliaGPU/AMDGPU.jl", rev="master")
-      Pkg.instantiate()
-      include("test/gpu/amd.jl")'
-    timeout_in_minutes: 30
+  # - label: "AMD GPUs -- AMDGPU.jl"
+  #   plugins:
+  #     - JuliaCI/julia#v1:
+  #         version: 1.8
+  #   agents:
+  #     queue: "juliagpu"
+  #     rocm: "*"
+  #     rocmgpu: "*"
+  #   env:
+  #     JULIA_AMDGPU_CORE_MUST_LOAD: "1"
+  #     JULIA_AMDGPU_HIP_MUST_LOAD: "1"
+  #   command: |
+  #     julia --color=yes --project -e '
+  #     using Pkg
+  #     Pkg.add("AMDGPU")
+  #     Pkg.instantiate()
+  #     include("test/gpu/amd.jl")'
+  #   timeout_in_minutes: 30
 
   - label: "Intel GPUs -- oneAPI.jl"
     plugins:
