+++ src/gpmr.jl
@@ -145,8 +145,8 @@ function gpmr!(solver :: GpmrSolver{T,FC,S}, A, B, b :: AbstractVector{FC}, c ::
   m, n = size(A)
   s, t = size(B)
   (m == solver.m && n == solver.n) || error("(solver.m, solver.n) = ($(solver.m), $(solver.n)) is inconsistent with size(A) = ($m, $n)")
-  ixm = kaxe(b)
-  ixn = kaxe(c)
+  ixm = kaxis(solver.x)
+  ixn = kaxis(solver.y)
   m == t         || error("Inconsistent problem size")
   s == n         || error("Inconsistent problem size")
   length(b) == m || error("Inconsistent problem size")
