### 2023-02-13 73dfe2f
Update the docstring of kaxes
