+++ docs/src/reference.md
@@ -11,7 +11,6 @@ Krylov.FloatOrComplex
 Krylov.niterations
 Krylov.Aprod
 Krylov.Atprod
-Krylov.kaxis
 Krylov.kaxes
 Base.show
 ```
