### 2023-02-13 0670e86
Return an error if S an AbstractSparseVector
