+++ src/krylov_solvers.jl
@@ -76,6 +76,7 @@ mutable struct MinresSolver{T,FC,S} <: KrylovSolver{T,FC,S}
 end
 
 function MinresSolver(m, n, ::Type{S}; ixm=m, ixn=n, window :: Int=5) where S <: AbstractVector
+  S <: AbstractSparseVector && error("The storage type S can't be a sparse vector for performance reasons")
   dense = S <: DenseVector
   FC = eltype(S)
   T  = real(FC)
@@ -124,6 +125,7 @@ mutable struct CgSolver{T,FC,S} <: KrylovSolver{T,FC,S}
 end
 
 function CgSolver(m, n, ::Type{S}; ixm=m, ixn=n) where S <: AbstractVector
+  S <: AbstractSparseVector && error("The storage type S can't be a sparse vector for performance reasons")
   dense = S <: DenseVector
   FC = eltype(S)
   T  = real(FC)
@@ -170,6 +172,7 @@ mutable struct CrSolver{T,FC,S} <: KrylovSolver{T,FC,S}
 end
 
 function CrSolver(m, n, ::Type{S}; ixm=m, ixn=n) where S <: AbstractVector
+  S <: AbstractSparseVector && error("The storage type S can't be a sparse vector for performance reasons")
   dense = S <: DenseVector
   FC = eltype(S)
   T  = real(FC)
@@ -220,6 +223,7 @@ mutable struct SymmlqSolver{T,FC,S} <: KrylovSolver{T,FC,S}
 end
 
 function SymmlqSolver(m, n, ::Type{S}; ixm=m, ixn=n, window :: Int=5) where S <: AbstractVector
+  S <: AbstractSparseVector && error("The storage type S can't be a sparse vector for performance reasons")
   dense = S <: DenseVector
   FC      = eltype(S)
   T       = real(FC)
@@ -270,6 +274,7 @@ mutable struct CgLanczosSolver{T,FC,S} <: KrylovSolver{T,FC,S}
 end
 
 function CgLanczosSolver(m, n, ::Type{S}; ixm=m, ixn=n) where S <: AbstractVector
+  S <: AbstractSparseVector && error("The storage type S can't be a sparse vector for performance reasons")
   dense = S <: DenseVector
   FC      = eltype(S)
   T       = real(FC)
@@ -323,6 +328,7 @@ mutable struct CgLanczosShiftSolver{T,FC,S} <: KrylovSolver{T,FC,S}
 end
 
 function CgLanczosShiftSolver(m, n, nshifts, ::Type{S}; ixm=m, ixn=n) where S <: AbstractVector
+  S <: AbstractSparseVector && error("The storage type S can't be a sparse vector for performance reasons")
   dense = S <: DenseVector
   FC         = eltype(S)
   T          = real(FC)
@@ -378,6 +384,7 @@ mutable struct MinresQlpSolver{T,FC,S} <: KrylovSolver{T,FC,S}
 end
 
 function MinresQlpSolver(m, n, ::Type{S}; ixm=m, ixn=n) where S <: AbstractVector
+  S <: AbstractSparseVector && error("The storage type S can't be a sparse vector for performance reasons")
   dense = S <: DenseVector
   FC      = eltype(S)
   T       = real(FC)
@@ -430,6 +437,7 @@ mutable struct DqgmresSolver{T,FC,S} <: KrylovSolver{T,FC,S}
 end
 
 function DqgmresSolver(m, n, memory, ::Type{S}; ixm=m, ixn=n) where S <: AbstractVector
+  S <: AbstractSparseVector && error("The storage type S can't be a sparse vector for performance reasons")
   dense = S <: DenseVector
[…]
