### 2023-02-13 68d9f43
Apply suggestions from code review
Co-authored-by: tmigot <tangi.migot@gmail.com>
