### 2023-02-15 cc175bc
Update example with Nvidia IC0 preconditioner
