### 2023-02-16 5043bb2
[documentation] Update the page related to the GPU support
