### 2023-02-18 #4970 @amontoison
@jedbrown
I checked again [this list of people](https://bit.ly/joss-reviewers) and I have new suggestions:
`@PetrKryslUCSD`, `@ChrisRackauckas`, `@vchuravy`, `@ysimillides`, `@jgoldfar`, `@williamfgc`. `@mlxd`.
