### 2023-02-20 1c24cf5
Update docs/src/gpu.md
Co-authored-by: Paul Raynaud <paul.raynaud66@hotmail.fr>
