### 2023-02-23 #4970 @jedbrown
I appreciate your patience. I would love for things to move faster, but people are massively oversubscribed and it tends to take many requests per reviewer acceptance. This situation is not unique to JOSS -- I have submitted manuscripts that are nearly a year out and every editor I know encounters this -- though we try to move things along more quickly via the review structure at JOSS. There was a structural discussion in the pre-review and the paper was updated (thanks!). I'm working on finding reviewers and will update as soon as possible.
