### 2023-02-26 #4970 @jedbrown
Thanks for your comment regarding PETSc. The old table had been stale (I think nobody intended it to be comprehensive, but obviously people were looking at it as though it was). You can see [this updated table](https://petsc.org/main/overview/linear_solve_table/#krylov-methods) for a current list. Many of the flexible and pipelined methods generate the same Krylov space (if run in exact arithmetic with a linear preconditioner), albeit with different parallel and computational semantics.
