### 2023-02-26 #5187 @editorialbot
[REVIEW]: Krylov.jl: A Julia basket of hand-picked Krylov methods
**Submitting author:** @amontoison (<a href="http://orcid.org/0000-0002-3403-5450">Alexis Montoison</a>)
**Repository:** https://github.com/JuliaSmoothOptimizers/Krylov.jl
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.9.4
