### 2023-03-06 #5187 @prj-
Hello @jedbrown and @amontoison, here is the current state of my checklist, with the remaining unticked checkboxes that should be addressed if possible.

Thanks.

> ### General checks
> * [ ]  **License:** Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?

The file `LICENSE.md` has a leading `Copyright (c) 2015-2019: Dominique Orban`. It seems outdated, should this be updated? If not, what's copyrighted?

> * [ ]  **Reproducibility:** If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.

I don't think the two provided examples enforce a check with respect to a reference solution. The authors do not state as well what should be the solution. This is very minor, but it would be nice to enforce such a check or at least add a comment with the expected solution.

> ### Functionality
> * [ ]  **Functionality:** Have the functional claims of the software been confirmed?

The second code snippet does not run when CUDA is not available on the system. Could the authors add some minimal error checking (or even better, add a switch to a more conventional CPU backend) to avoid unhandled runtime exceptions?

> ### Software paper
> * [ ]  **State of the field:** Do the authors describe how this software compares to other commonly-used packages?

Line 22, "Largest collection of Krylov processes and methods", does not seem very relevant to me. First of all, why should quantity matter over quality? Second, though it should certainly be stated that the collection is very complete, it would be more interesting, in my opinion, to justify the use of this package over other Julia alternatives such as KrylovKit, KrylovMethods, IterativeSolvers, and such.

### Other minor points
* [ ] Line 35, missing `if`.
* [ ] `Nvidia` should be typeset `NVIDIA` throughout the document.
* [ ] Line 53, missing `,` before `and Intel GPUs`.
* [ ] Line 72, `Operator-vector products and vector operations are the most expensive operations in Krylov.jl. We rely on BLAS routines as much as possible to perform those operations`. This statement is not clear to me, isn't the user supposed to provide the operator-vector product, in which case, how can you certify that you are relying on BLAS there as well? I get what you mean, but please rephrase.
* [ ] Line 92, `the Cholesky factor to model`, maybe best rephrased as `the Cholesky factor to approximate`.
