+++ paper/paper.md
@@ -4,6 +4,7 @@ tags:
   - Julia
   - linear algebra
   - Krylov methods
+  - Krylov processes
   - sparse linear systems
   - GPU computing
 authors:
@@ -16,7 +17,7 @@ authors:
 affiliations:
  - name: GERAD and Department of Mathematics and Industrial Engineering, Polytechnique Montréal, QC, Canada.
    index: 1
-date: 15 February 2023
+date: 6 March 2023
 bibliography: paper.bib
 header-includes: |
   \usepackage{booktabs}
@@ -54,14 +55,15 @@ We refer interested readers to [@ipsen-meyer-1998] for an introduction to Krylov
 
 ##  Largest collection of Krylov processes and methods
 
-Krylov.jl aims to provide a unified interface for the largest collection of Krylov processes and methods, all programming languages taken together, with six and thirty-three implementations, respectively:
+Krylov.jl aims to provide a user-friendly and unified interface for the largest collection of Krylov processes and methods, all programming languages taken together, with six and thirty-three implementations, respectively:
 
 - \textbf{Krylov processes}: \textsc{Arnoldi}, \textsc{Golub-Kahan}, \textsc{Hermitian Lanczos}, \textsc{Montoison-Orban}, \textsc{Non-Hermitian Lanczos},  \textsc{Saunders-Simon-Yip};
 - \textbf{Krylov methods}: \textsc{Bicgstab}, \textsc{Bilq}, \textsc{Bilqr}, \textsc{Cg}, \textsc{Cg-lanczos}, \textsc{Cg-lanczos-shift}, \textsc{Cgls}, \textsc{Cgne}, \textsc{Cgs}, \textsc{Cr}, \textsc{Craig}, \textsc{Craigmr}, \textsc{Crls}, \textsc{Crmr}, \textsc{Diom}, \textsc{Dqgmres}, \textsc{Fgmres}, \textsc{Fom}, \textsc{Gmres}, \textsc{Gpmr}, \textsc{Lnlq}, \textsc{Lslq}, \textsc{Lsmr}, \textsc{Lsqr}, \textsc{Minres}, \textsc{Minres-qlp}, \textsc{Qmr}, \textsc{Symmlq}, \textsc{Tricg}, \textsc{Trilqr}, \textsc{Trimr}, \textsc{Usymlq}, \textsc{Usymqr}.
 
+Hence Krylov.jl is a suitable toolbox for easily comparing existing methods with each other as well as new ones.
 MATLAB [@MATLAB] and PETSc [@petsc] have eleven and eighteen distinct Krylov methods, respectively.
 Note that we only consider the number of Krylov methods that generate different iterates without preconditioning.
-Variants with preconditioning are not counted except it is a flexible one such as \textsc{Fgmres}.
+Variants with preconditioning are not counted except if it is a flexible one such as \textsc{Fgmres}.
 
 Some processes and methods are not available elsewhere and are the product of our own research.
 References for each process and method are available in the extensive [documentation](https://juliasmoothoptimizers.github.io/Krylov.jl/stable/).
@@ -73,13 +75,13 @@ Although most personal computers offer IEEE 754 single and double precision comp
 In addition, software libraries such as the GNU MPFR, shipped with Julia, let users experiment with computations in variable, extended precision at the software level with the `BigFloat` data type.
 Working in high precision has obvious benefits in terms of accuracy.
 
-## Support for Nvidia, AMD and Intel GPUs
+## Support for NVIDIA, AMD and Intel GPUs
 
[…]
