### 2023-03-06 28227d1
[documentation] Update the GPU section
