+++ docs/src/gpu.md
@@ -14,16 +14,18 @@ Problems stored in CPU format (`Matrix` and `Vector`) must first be converted to
 ```julia
 using CUDA, Krylov
 
-# CPU Arrays
-A_cpu = rand(20, 20)
-b_cpu = rand(20)
+if CUDA.functional()
+  # CPU Arrays
+  A_cpu = rand(20, 20)
+  b_cpu = rand(20)
 
-# GPU Arrays
-A_gpu = CuMatrix(A_cpu)
-b_gpu = CuVector(b_cpu)
+  # GPU Arrays
+  A_gpu = CuMatrix(A_cpu)
+  b_gpu = CuVector(b_cpu)
 
-# Solve a square and dense system on an Nivida GPU
-x, stats = bilq(A_gpu, b_gpu)
+  # Solve a square and dense system on an Nivida GPU
+  x, stats = bilq(A_gpu, b_gpu)
+end
 ```
 
 Sparse matrices have a specific storage on Nvidia GPUs (`CuSparseMatrixCSC`, `CuSparseMatrixCSR` or `CuSparseMatrixCOO`):
@@ -32,16 +34,18 @@ Sparse matrices have a specific storage on Nvidia GPUs (`CuSparseMatrixCSC`, `Cu
 using CUDA, Krylov
 using CUDA.CUSPARSE, SparseArrays
 
-# CPU Arrays
-A_cpu = sprand(200, 100, 0.3)
-b_cpu = rand(200)
+if CUDA.functional()
+  # CPU Arrays
+  A_cpu = sprand(200, 100, 0.3)
+  b_cpu = rand(200)
 
-# GPU Arrays
-A_gpu = CuSparseMatrixCSC(A_cpu)
-b_gpu = CuVector(b_cpu)
+  # GPU Arrays
+  A_gpu = CuSparseMatrixCSC(A_cpu)
+  b_gpu = CuVector(b_cpu)
 
-# Solve a rectangular and sparse system on an Nvidia GPU
-x, stats = lsmr(A_gpu, b_gpu)
+  # Solve a rectangular and sparse system on an Nvidia GPU
+  x, stats = lsmr(A_gpu, b_gpu)
+end
 ```
 
 Optimized operator-vector products that exploit GPU features can be also used by means of linear operators.
@@ -55,37 +59,39 @@ can be applied directly on GPU thanks to efficient operators that take advantage
 using SparseArrays, Krylov, LinearOperators
 using CUDA, CUDA.CUSPARSE
 
-# Transfer the linear system from the CPU to the GPU
-A_gpu = CuSparseMatrixCSR(A_cpu)  # A_gpu = CuSparseMatrixCSC(A_cpu)
-b_gpu = CuVector(b_cpu)
+if CUDA.functional()
+  # Transfer the linear system from the CPU to the GPU
+  A_gpu = CuSparseMatrixCSR(A_cpu)  # A_gpu = CuSparseMatrixCSC(A_cpu)
+  b_gpu = CuVector(b_cpu)
 
-# Incomplete decomposition LLᴴ ≈ A for CuSparseMatrixCSC or CuSparseMatrixCSR matrices
-P = ic02(A_gpu, 'O')
+  # Incomplete decomposition LLᴴ ≈ A for CuSparseMatrixCSC or CuSparseMatrixCSR matrices
+  P = ic02(A_gpu, 'O')
 
-# Additional vector required for solving triangular systems
-n = length(b_gpu)
-T = eltype(b_gpu)
-z = similar(CuVector{T}, n)
+  # Additional vector required for solving triangular systems
+  n = length(b_gpu)
+  T = eltype(b_gpu)
+  z = similar(CuVector{T}, n)
 
-# Solve Py = x
[…]
