### 2023-03-08 662176f
Update paper/paper.md
Co-authored-by: Dominique <dominique.orban@gmail.com>
