### 2023-03-10 aaf83e5
Update the reference of the GPMR paper
