### 2023-03-24 088e40c
Fix the documentation for CUDA.jl 4.1
