### 2023-03-24 fcbf189
Fix the documentation for CUDA.jl 4.1
