### 2023-03-24 a8a1213
Update docs/src/gpu.md
Co-authored-by: tmigot <tangi.migot@gmail.com>
