+++ .github/codecov.yml
@@ -1,0 +1,8 @@
+coverage:
+  status:
+    project:
+      default:
+        # Drops on the order 0.01% are typical even when no change occurs
+        # Having this threshold set a little higher (0.5%) than that makes it 
+        # a little more tolerant to fluctuations
+        threshold: 0.5%
