### 2023-03-25 265296f
Update CI scripts for the JOSS paper
