### 2023-03-25 c61864b
Update CI scripts for the JOSS paper
