### 2023-03-25 1ff856d
Update CI scripts for the JOSS paper
