### 2023-03-25 83a5b86
Update CI scripts for the JOSS paper
