### 2023-03-25 4e47644
Update CI scripts for the JOSS paper
