+++ paper/paper.md
@@ -17,12 +17,12 @@ authors:
 affiliations:
  - name: GERAD and Department of Mathematics and Industrial Engineering, Polytechnique Montréal, QC, Canada.
    index: 1
-date: 6 March 2023
+date: 25 March 2023
 bibliography: paper.bib
 header-includes: |
   \usepackage{booktabs}
   \usepackage{fontspec}
-  \setmonofont[Path = ./, Scale=0.7]{JuliaMono-Regular.ttf}
+  \setmonofont[Path = ./, Scale=0.68]{JuliaMono-Regular.ttf}
 ---
 
 # Summary
@@ -61,12 +61,14 @@ Krylov.jl aims to provide a user-friendly and unified interface for the largest 
 - \textbf{Krylov methods}: \textsc{Bicgstab}, \textsc{Bilq}, \textsc{Bilqr}, \textsc{Cg}, \textsc{Cg-lanczos}, \textsc{Cg-lanczos-shift}, \textsc{Cgls}, \textsc{Cgne}, \textsc{Cgs}, \textsc{Cr}, \textsc{Craig}, \textsc{Craigmr}, \textsc{Crls}, \textsc{Crmr}, \textsc{Diom}, \textsc{Dqgmres}, \textsc{Fgmres}, \textsc{Fom}, \textsc{Gmres}, \textsc{Gpmr}, \textsc{Lnlq}, \textsc{Lslq}, \textsc{Lsmr}, \textsc{Lsqr}, \textsc{Minres}, \textsc{Minres-qlp}, \textsc{Qmr}, \textsc{Symmlq}, \textsc{Tricg}, \textsc{Trilqr}, \textsc{Trimr}, \textsc{Usymlq}, \textsc{Usymqr}.
 
 Hence Krylov.jl is a suitable toolbox for easily comparing existing methods with each other as well as new ones.
-MATLAB [@MATLAB] and PETSc [@petsc] have eleven and eighteen distinct Krylov methods, respectively.
+The number of distinct Krylov methods is eighteen for PETSc [@petsc], eleven for MATLAB @MATLAB and [KrylovMethods.jl](https://github.com/JuliaInv/KrylovMethods.jl), nine for [IterativeSolvers.jl](https://github.com/JuliaLinearAlgebra/IterativeSolvers.jl) and three for [KrylovKit.jl](https://github.com/Jutho/KrylovKit.jl).
+However Krylov.jl doesn't have implementations of Block-Krylov methods unlike some alternatives.
 Note that we only consider the number of Krylov methods that generate different iterates without preconditioning.
 Variants with preconditioning are not counted except if it is a flexible one such as \textsc{Fgmres}.
 
 Some processes and methods are not available elsewhere and are the product of our own research.
 References for each process and method are available in the extensive [documentation](https://juliasmoothoptimizers.github.io/Krylov.jl/stable/).
+Beyond the number of methods, Krylov.jl is the only package that offers all of the features that we are going to describe.
 
 ## Support for any floating-point system supported by Julia
 
@@ -101,7 +103,9 @@ In-place methods limit memory allocations and deallocations, which are particula
 ## Performance optimizations and storage requirements
 
 Operator-vector products and vector operations are the most expensive operations in Krylov.jl.
-We rely on BLAS routines as much as possible to perform those operations.
+To speed up theses procedures, we rely on BLAS when the linear problems are solved with CPUs and stored in a precision supported by BLAS.
[…]
