### 2023-03-25 #5187 @amontoison
@prj-, I worked on a last PR yesterday and today to adress all your comments.

> The file `LICENSE.md` has a leading `Copyright (c) 2015-2019: Dominique Orban`. It seems outdated, should this be updated? If not, what's copyrighted?

Thanks, I updated the LICENSE.md file. (https://github.com/JuliaSmoothOptimizers/Krylov.jl/pull/712)

> I don't think the two provided examples enforce a check with respect to a reference solution. The authors do not state as well what should be the solution. This is very minor, but it would be nice to enforce such a check or at least add a comment with the expected solution.

For the reproducibility, I pushed the example files on the branch `joss-paper`. Each example is in a folder `paper/exampleX` with `Project.toml` and `Manifest.toml` files.
With a `julia --project` you can start Julia with exactly the same environnement that we used to run the examples.
I also updated the CI script `.buildkite/pipeline.yml` to run the example files to check that we have don't have errors and we can run the examples with another computer. (see [CI builds](https://buildkite.com/julialang/krylov-dot-jl/builds?branch=amontoison%3Ajoss-paper))
Do you want an additional check to ensure that we have an expected solution?

> The second code snippet does not run when CUDA is not available on the system. Could the authors add some minimal error checking (or even better, add a switch to a more conventional CPU backend) to avoid unhandled runtime exceptions?

I added the following test in the second code snippet to ensure that we have the CUDA drivers a NVIDIA GPU available:
```julia
if CUDA.functional()
  ...
end
```
I also updated the documentation (https://github.com/JuliaSmoothOptimizers/Krylov.jl/pull/714) to be sure that users don't try to store and solve linear problems on GPUs if they don't have the hardware associated.

> Line 22, "Largest collection of Krylov processes and methods", does not seem very relevant to me. First of all, why should quantity matter over quality? Second, though it should certainly be stated that the collection is very complete, it would be more interesting, in my opinion, to justify the use of this package over other Julia alternatives such as KrylovKit, KrylovMethods, IterativeSolvers, and such.

Krylov.jl represents my thesis work and I see the package as a toolbox of Krylov processes and methods.
As a researcher of new Krylov methods and Krylov processes, I find it's practical for benchmarking and comparing the behavior of various Krylov subpaces. It's also great for teaching.
The paragraph "Largest collection of Krylov processes and methods" was intended to highlight this characteristic.
I reformulated the parapraph and also mentioned the packages KrylovKit.jl, KrylovMethods.jl and IterativeSolvers.jl.
With the other paragraphs, we try to justify as best we can that we have quantity and quality.

> ### Other minor points
>     * Line 35, missing `if`.
>     * `Nvidia` should be typeset `NVIDIA` throughout the document.
>     * Line 53, missing `,` before `and Intel GPUs`.
>     * Line 92, `the Cholesky factor to model`, maybe best rephrased as `the Cholesky factor to approximate`.

I fixed these typos in the PR https://github.com/JuliaSmoothOptimizers/Krylov.jl/pull/712.

>     * Line 72, `Operator-vector products and vector operations are the most expensive operations in Krylov.jl. We rely on BLAS routines as much as possible to perform those operations`. This statement is not clear to me, isn't the user supposed to provide the operator-vector product, in which case, how can you certify that you are relying on BLAS there as well? I get what you mean, but please rephrase.

I clarified and rephrased this sentence in the PR https://github.com/JuliaSmoothOptimizers/Krylov.jl/pull/720.

I will regenerate the PDF of the paper here as soon as the last PR is merged.
