### 2023-03-26 c23b07d
Apply suggestions from code review
Co-authored-by: Dominique <dominique.orban@gmail.com>
