### 2023-03-31 19542fa
[documentation] Use CUDA.zeros instead of similar
