### 2023-03-31 a4f01ac
Update examples of the JOSS paper
