+++ paper/example2/example2.jl
@@ -1,5 +1,6 @@
 using LinearAlgebra                # Linear algebra library of Julia
 using SparseArrays                 # Sparse library of Julia
+using Test                         # Test library of Julia
 using Krylov                       # Krylov methods and processes
 using LinearOperators              # Linear operators
 using MatrixMarket                 # Reader of matrices stored in the Matrix Market format
@@ -8,23 +9,26 @@ using CUDA                         # Interface to NVIDIA GPUs
 using CUDA.CUSPARSE                # NVIDIA CUSPARSE library
 
 if CUDA.functional()
-  ssmc = ssmc_db()
-  matrices = ssmc_matrices(ssmc, "Bai", "mhd1280b")
+  ssmc = ssmc_db(verbose=false)
+  matrices = ssmc_matrices(ssmc, "Sinclair", "3Dspectralwave2")
   paths = fetch_ssmc(matrices, format="MM")
-  path_A = joinpath(paths[1], "mhd1280b.mtx")
-  A_cpu = MatrixMarket.mmread(path_A)
+  path_A = joinpath(paths[1], "3Dspectralwave2.mtx")
+
+  # A is an Hermitian and positive definite matrix of size 292008 x 292008
+  A_cpu = MatrixMarket.mmread(path_A) + 50I
   m, n = size(A_cpu)
-  b_cpu = ones(ComplexF64, m)
+  xstar = ones(ComplexF64, m)
+  b_cpu = A_cpu * xstar
 
   # Transfer the linear system from the CPU to the GPU
   A_gpu = CuSparseMatrixCSR(A_cpu)
   b_gpu = CuVector(b_cpu)
 
-  # Incomplete Cholesky decomposition LLᴴ ≈ A with zero fill-in
-  P = ic02(A_gpu, 'O')
+  # Incomplete Cholesky factorization LLᴴ ≈ A with zero fill-in
+  P = ic02(A_gpu)
 
   # Additional vector required for solving triangular systems
-  z = similar(CuVector{ComplexF64}, n)
+  z = CUDA.zeros(ComplexF64, n)
 
   # Solve Py = x
   function ldiv_ic0!(P, x, y, z)
@@ -39,6 +43,10 @@ if CUDA.functional()
   hermitian = true
   P⁻¹ = LinearOperator(T, m, n, symmetric, hermitian, (y, x) -> ldiv_ic0!(P, x, y, z))
 
-  # Solve a Hermitian positive definite system with an incomplete Cholesky factorization preconditioner
-  x, stats = cg(A_gpu, b_gpu, M=P⁻¹)
+  # Solve an Hermitian positive definite system with an incomplete Cholesky factorization preconditioner
+  x_gpu, stats = cg(A_gpu, b_gpu, M=P⁻¹)
+
+  # Check the solution returned by the conjugate gradient method
+  x_cpu = Vector{ComplexF64}(x_gpu)
+  @test norm(x_cpu - xstar) ≤ 1e-5
 end
