### 2023-03-31 e1cd29a
Update examples of the JOSS paper
