+++ paper/paper.md
@@ -17,7 +17,7 @@ authors:
 affiliations:
  - name: GERAD and Department of Mathematics and Industrial Engineering, Polytechnique Montréal, QC, Canada.
    index: 1
-date: 25 March 2023
+date: 31 March 2023
 bibliography: paper.bib
 header-includes: |
   \usepackage{booktabs}
@@ -62,7 +62,7 @@ Krylov.jl aims to provide a user-friendly and unified interface for the largest 
 
 Hence Krylov.jl is a suitable toolbox for easily comparing existing methods with each other as well as new ones.
 The number of distinct Krylov methods is eighteen for PETSc [@petsc], eleven for @MATLAB and [KrylovMethods.jl](https://github.com/JuliaInv/KrylovMethods.jl), nine for [IterativeSolvers.jl](https://github.com/JuliaLinearAlgebra/IterativeSolvers.jl) and three for [KrylovKit.jl](https://github.com/Jutho/KrylovKit.jl).
-However Krylov.jl doesn't have implementations of Block-Krylov methods unlike some alternatives, except for special cases, including \textsc{Tricg}, \textsc{Trimr}, and \textsc{Gpmr}.
+However Krylov.jl doesn't have implementations of recycling Krylov methods nor block Krylov methods unlike some alternatives, except for special cases, including \textsc{Tricg}, \textsc{Trimr}, and \textsc{Gpmr}.
 Note that we only consider the number of Krylov methods that generate different iterates without preconditioning.
 Variants with preconditioning are not counted except if it is a flexible one such as \textsc{Fgmres}.
 
@@ -122,6 +122,7 @@ Another example based on a simplistic Newton method without linesearch for conve
 ```julia
 using LinearAlgebra    # Linear algebra library of Julia
 using SparseArrays     # Sparse library of Julia
+using Test             # Test library of Julia
 using Krylov           # Krylov methods and processes
 using LinearOperators  # Linear operators
 using ForwardDiff      # Automatic differentiation
@@ -153,24 +154,31 @@ end
 
 T = Float128  # IEEE quadruple precision
 x₀ = ones(T, 2)
-F(x) = [x[1]^4 - 3; exp(x[2]) - 2; log(x[1]) - x[2]^2]         # F(x)
+t = T[1, 2, 3, 4, 5, 6, 7, 8]
+y = T[10.272, 13.190, 16.936, 21.746, 27.923, 35.854, 46.037, 59.112]
+F(x) = [x[1] * exp(x[2] * t[i]) - y[i] for i=1:8]              # F(x)
 J(y, x, v) = ForwardDiff.derivative!(y, t -> F(x + t * v), 0)  # y ← JF(x)v
 Jᵀ(y, x, w) = ForwardDiff.gradient!(y, x -> dot(F(x), w), x)   # y ← JFᵀ(x)w
 symmetric = hermitian = false
-JF(x) = LinearOperator(T, 3, 2, symmetric, hermitian, (y, v) -> J(y, x, v),   # non-transpose
+JF(x) = LinearOperator(T, 8, 2, symmetric, hermitian, (y, v) -> J(y, x, v),   # non-transpose
                                                       (y, w) -> Jᵀ(y, x, w),  # transpose
                                                       (y, w) -> Jᵀ(y, x, w))  # conjugate transpose
-xstar = gauss_newton(F, JF, x₀)
+x = gauss_newton(F, JF, x₀)
+
+# Check the solution returned by the Gauss-Newton method
+xstar = T[8, 0.25]
+@test norm(x - xstar) ≤ 1e-4
 ```
 
[…]
