### 2023-03-31 e6b6ff1
Update examples of the JOSS paper
