+++ paper/example1/Project.toml
@@ -6,3 +6,4 @@ LinearOperators = "5c8ed15e-5a4c-59e4-a42b-c7e8811fb125"
 MKL = "33e6dc65-8f57-5167-99aa-e5a354878fb2"
 Quadmath = "be4d8f0f-7fa4-5f49-b795-2f01399ab2dd"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
