### 2023-03-31 aff2922
Update examples of the JOSS paper
