### 2023-03-31 b6b3a9d
Update examples of the JOSS paper
