+++ paper/example1/example1.jl
@@ -31,11 +31,12 @@ function gauss_newton(F, JF, x₀::AbstractVector{T}; itmax = 200, tol = √eps(
 end
 
 T = Float128  # IEEE quadruple precision
+x_exact = T[8, 0.25]
 x₀ = ones(T, 2)
 t = T[1, 2, 3, 4, 5, 6, 7, 8]
-y = T[10.272, 13.190, 16.936, 21.746, 27.923, 35.854, 46.037, 59.112]
+y = [trunc(x_exact[1] * exp(x_exact[2] * t[i]), digits=3) for i=1:8]
 F(x) = [x[1] * exp(x[2] * t[i]) - y[i] for i=1:8]              # F(x)
-J(y, x, v) = ForwardDiff.derivative!(y, t -> F(x + t * v), 0)  # y ← JF(x)v
+J(y, x, v) = ForwardDiff.derivative!(y, h -> F(x + h * v), 0)  # y ← JF(x)v
 Jᵀ(y, x, w) = ForwardDiff.gradient!(y, x -> dot(F(x), w), x)   # y ← JFᵀ(x)w
 symmetric = hermitian = false
 JF(x) = LinearOperator(T, 8, 2, symmetric, hermitian, (y, v) -> J(y, x, v),   # non-transpose
@@ -44,5 +45,4 @@ JF(x) = LinearOperator(T, 8, 2, symmetric, hermitian, (y, v) -> J(y, x, v),   # 
 x = gauss_newton(F, JF, x₀)
 
 # Check the solution returned by the Gauss-Newton method
-xstar = T[8, 0.25]
-@test norm(x - xstar) ≤ 1e-4
+@test norm(x - x_exact) ≤ 1e-4
