### 2023-04-03 #5187 @prj-
Looks good, thanks. There are still a couple of typos, line 97, `"In-place methods"` (extra `s`) and line 105, `including Sʏᴍᴍʟᴏ̨, Cʀ, and Mɪɴʀᴇs` (missing `,`), but you can probably wait for the comments of the other reviewer before generating a new proof, it's up to you. I'll have a last look then.
