+++ src/diom.jl
@@ -15,7 +15,7 @@ export diom, diom!
                       memory::Int=20, M=I, N=I, ldiv::Bool=false,
                       reorthogonalization::Bool=false, atol::T=√eps(T),
                       rtol::T=√eps(T), itmax::Int=0,
-                      verbose::Int=0, history::Bool=false,
+                      timemax::Float64=Inf, verbose::Int=0, history::Bool=false,
                       callback=solver->false, iostream::IO=kstdout)
 
 `T` is an `AbstractFloat` such as `Float32`, `Float64` or `BigFloat`.
@@ -54,6 +54,7 @@ and indefinite systems of linear equations can be handled by this single algorit
 * `atol`: absolute stopping tolerance based on the residual norm;
 * `rtol`: relative stopping tolerance based on the residual norm;
 * `itmax`: the maximum number of iterations. If `itmax=0`, the default number of iterations is set to `2n`;
+* `timemax`: the time limit in seconds;
 * `verbose`: additional details can be displayed if verbose mode is enabled (verbose > 0). Information will be displayed every `verbose` iterations;
 * `history`: collect additional statistics on the run such as residual norms, or Aᴴ-residual norms;
 * `callback`: function or functor called as `callback(solver)` that returns `true` if the Krylov method should terminate, and `false` otherwise;
@@ -105,9 +106,10 @@ function diom!(solver :: DiomSolver{T,FC,S}, A, b :: AbstractVector{FC};
                M=I, N=I, ldiv :: Bool=false,
                reorthogonalization :: Bool=false, atol :: T=√eps(T),
                rtol :: T=√eps(T), itmax :: Int=0,
-               verbose :: Int=0, history :: Bool=false,
+               timemax :: Float64=Inf, verbose :: Int=0, history :: Bool=false,
                callback = solver -> false, iostream :: IO=kstdout) where {T <: AbstractFloat, FC <: FloatOrComplex{T}, S <: AbstractVector{FC}}
 
+  start_time = time()
   m, n = size(A)
   (m == solver.m && n == solver.n) || error("(solver.m, solver.n) = ($(solver.m), $(solver.n)) is inconsistent with size(A) = ($m, $n)")
   m == n || error("System must be square")
@@ -182,8 +184,9 @@ function diom!(solver :: DiomSolver{T,FC,S}, A, b :: AbstractVector{FC};
   tired = iter ≥ itmax
   status = "unknown"
   user_requested_exit = false
+  overtimed = false
 
-  while !(solved || tired || user_requested_exit)
+  while !(solved || tired || user_requested_exit || overtimed)
 
     # Update iteration index.
     iter = iter + 1
@@ -282,12 +285,16 @@ function diom!(solver :: DiomSolver{T,FC,S}, A, b :: AbstractVector{FC};
     resid_decrease_lim = rNorm ≤ ε
     solved = resid_decrease_lim || resid_decrease_mach
     tired = iter ≥ itmax
+    overtimed = time() - start_time > timemax
     kdisplay(iter, verbose) && @printf(iostream, "%5d  %7.1e\n", iter, rNorm)
   end
   (verbose > 0) && @printf(iostream, "\n")
+
+  # Termination status
   tired               && (status = "maximum number of iterations exceeded")
   solved              && (status = "solution good enough given atol and rtol")
   user_requested_exit && (status = "user-requested exit")
+  overtimed           && (status = "time limit exceeded")
 
   # Update x
   warm_start && @kaxpy!(n, one(FC), Δx, x)
