+++ src/symmlq.jl
@@ -18,7 +18,7 @@ export symmlq, symmlq!
                         λest::T=zero(T), etol::T=√eps(T),
                         conlim::T=1/√eps(T), atol::T=√eps(T),
                         rtol::T=√eps(T), itmax::Int=0,
-                        verbose::Int=0, history::Bool=false,
+                        timemax::Float64=Inf, verbose::Int=0, history::Bool=false,
                         callback=solver->false, iostream::IO=kstdout)
 
 `T` is an `AbstractFloat` such as `Float32`, `Float64` or `BigFloat`.
@@ -58,6 +58,7 @@ SYMMLQ produces monotonic errors ‖x* - x‖₂.
 * `atol`: absolute stopping tolerance based on the residual norm;
 * `rtol`: relative stopping tolerance based on the residual norm;
 * `itmax`: the maximum number of iterations. If `itmax=0`, the default number of iterations is set to `2n`;
+* `timemax`: the time limit in seconds;
 * `verbose`: additional details can be displayed if verbose mode is enabled (verbose > 0). Information will be displayed every `verbose` iterations;
 * `history`: collect additional statistics on the run such as residual norms, or Aᴴ-residual norms;
 * `callback`: function or functor called as `callback(solver)` that returns `true` if the Krylov method should terminate, and `false` otherwise;
@@ -108,9 +109,10 @@ function symmlq!(solver :: SymmlqSolver{T,FC,S}, A, b :: AbstractVector{FC};
                  λest :: T=zero(T), etol :: T=√eps(T),
                  conlim :: T=1/√eps(T), atol :: T=√eps(T),
                  rtol :: T=√eps(T), itmax :: Int=0,
-                 verbose :: Int=0, history :: Bool=false,
+                 timemax :: Float64=Inf, verbose :: Int=0, history :: Bool=false,
                  callback = solver -> false, iostream :: IO=kstdout) where {T <: AbstractFloat, FC <: FloatOrComplex{T}, S <: AbstractVector{FC}}
 
+  start_time = time()
   m, n = size(A)
   (m == solver.m && n == solver.n) || error("(solver.m, solver.n) = ($(solver.m), $(solver.n)) is inconsistent with size(A) = ($m, $n)")
   m == n || error("System must be square")
@@ -252,8 +254,9 @@ function symmlq!(solver :: SymmlqSolver{T,FC,S}, A, b :: AbstractVector{FC};
   solved = zero_resid = solved_lq || solved_cg
   fwd_err = false
   user_requested_exit = false
+  overtimed = false
 
-  while ! (solved || tired || ill_cond || user_requested_exit)
+  while ! (solved || tired || ill_cond || user_requested_exit || overtimed)
     iter = iter + 1
 
     # Continue QR factorization
@@ -401,6 +404,7 @@ function symmlq!(solver :: SymmlqSolver{T,FC,S}, A, b :: AbstractVector{FC};
     zero_resid = solved_lq || solved_cg
     ill_cond = ill_cond_mach || ill_cond_lim
     solved = solved_mach || zero_resid || zero_resid_mach || zero_resid_lim || fwd_err || resid_decrease_mach
+    overtimed = time() - start_time > timemax
   end
   (verbose > 0) && @printf(iostream, "\n")
 
@@ -410,6 +414,7 @@ function symmlq!(solver :: SymmlqSolver{T,FC,S}, A, b :: AbstractVector{FC};
     @kaxpy!(m, ζbar, w̅, x)
   end
 
+  # Termination status
   tired               && (status = "maximum number of iterations exceeded")
   ill_cond_mach       && (status = "condition number seems too large for this machine")
[…]

[…]
