### 2023-05-05 e735826
Remove a test with a broken callback for FOM
