+++ test/test_fom.jl
@@ -126,13 +126,6 @@
       end
       
       # test callback function
-      solver = FomSolver(A, b)
-      tol = 1.0e-1
-      cb_n2 = TestCallbackN2(A, b, tol = tol)
-      fom!(solver, A, b, restart = true, callback = cb_n2)
-      @test solver.stats.status == "user-requested exit"
-      @test cb_n2(solver)
-
       @test_throws TypeError fom(A, b, restart = true, callback = solver -> "string", history = true)
     end
   end
