### 2023-05-05 e953d9e
Use a timer variable to store time() - start_time
