+++ src/cg.jl
@@ -222,12 +222,8 @@ function cg!(solver :: CgSolver{T,FC,S}, A, b :: AbstractVector{FC};
     iter = iter + 1
     tired = iter ≥ itmax
     user_requested_exit = callback(solver) :: Bool
-    println("iter: $iter")
-    println(timemax)
     timer = time() - start_time
-    println(timer)
     overtimed = timer > timemax
-    println(overtimed)
     kdisplay(iter, verbose) && @printf(iostream, "%5d  %7.1e  ", iter, rNorm)
   end
   (verbose > 0) && @printf(iostream, "\n")
