### 2023-05-06 a8eaa53
Test time_ns() instead of time()
