### 2023-05-08 e320603
[documentation] Remove the comment about unstable CUDA toolkits
