+++ src/cg_lanczos.jl
@@ -17,7 +17,7 @@ export cg_lanczos, cg_lanczos!
                             M=I, ldiv::Bool=false,
                             check_curvature::Bool=false, atol::T=√eps(T),
                             rtol::T=√eps(T), itmax::Int=0,
-                            verbose::Int=0, history::Bool=false,
+                            timemax::Float64=Inf, verbose::Int=0, history::Bool=false,
                             callback=solver->false, iostream::IO=kstdout)
 
 `T` is an `AbstractFloat` such as `Float32`, `Float64` or `BigFloat`.
@@ -49,6 +49,7 @@ The method does _not_ abort if A is not definite.
 * `atol`: absolute stopping tolerance based on the residual norm;
 * `rtol`: relative stopping tolerance based on the residual norm;
 * `itmax`: the maximum number of iterations. If `itmax=0`, the default number of iterations is set to `2n`;
+* `timemax`: the time limit in seconds;
 * `verbose`: additional details can be displayed if verbose mode is enabled (verbose > 0). Information will be displayed every `verbose` iterations;
 * `history`: collect additional statistics on the run such as residual norms, or Aᴴ-residual norms;
 * `callback`: function or functor called as `callback(solver)` that returns `true` if the Krylov method should terminate, and `false` otherwise;
@@ -98,9 +99,11 @@ function cg_lanczos!(solver :: CgLanczosSolver{T,FC,S}, A, b :: AbstractVector{F
                      M=I, ldiv :: Bool=false,
                      check_curvature :: Bool=false, atol :: T=√eps(T),
                      rtol :: T=√eps(T), itmax :: Int=0,
-                     verbose :: Int=0, history :: Bool=false,
+                     timemax :: Float64=Inf, verbose :: Int=0, history :: Bool=false,
                      callback = solver -> false, iostream :: IO=kstdout) where {T <: AbstractFloat, FC <: FloatOrComplex{T}, S <: AbstractVector{FC}}
 
+  start_time = time_ns()
+  timemax_ns = 1e9 * timemax
   m, n = size(A)
   (m == solver.m && n == solver.n) || error("(solver.m, solver.n) = ($(solver.m), $(solver.n)) is inconsistent with size(A) = ($m, $n)")
   m == n || error("System must be square")
@@ -172,9 +175,10 @@ function cg_lanczos!(solver :: CgLanczosSolver{T,FC,S}, A, b :: AbstractVector{F
   tired = iter ≥ itmax
   status = "unknown"
   user_requested_exit = false
+  overtimed = false
 
   # Main loop.
-  while ! (solved || tired || (check_curvature & indefinite) || user_requested_exit)
+  while ! (solved || tired || (check_curvature & indefinite) || user_requested_exit || overtimed)
     # Form next Lanczos vector.
     # βₖ₊₁Mvₖ₊₁ = Avₖ - δₖMvₖ - βₖMvₖ₋₁
     mul!(Mv_next, A, v)        # Mvₖ₊₁ ← Avₖ
@@ -218,13 +222,17 @@ function cg_lanczos!(solver :: CgLanczosSolver{T,FC,S}, A, b :: AbstractVector{F
     resid_decrease_lim = rNorm ≤ ε
     solved = resid_decrease_lim || resid_decrease_mach
     tired = iter ≥ itmax
+    timer = time_ns() - start_time
+    overtimed = timer > timemax_ns
   end
   (verbose > 0) && @printf(iostream, "\n")
 
+  # Termination status
   tired                          && (status = "maximum number of iterations exceeded")
   (check_curvature & indefinite) && (status = "negative curvature")
   solved                         && (status = "solution good enough given atol and rtol")
[…]

[…]
