### 2023-05-09 b87e4d6
Use metaprogramming to define the kwargs of all Krylov methods
