+++ src/trilqr.jl
@@ -69,18 +69,6 @@ USYMQR is used for solving dual system `Aᴴy = c` of size n × m.
 """
 function trilqr end
 
-function trilqr(A, b :: AbstractVector{FC}, c :: AbstractVector{FC}, x0 :: AbstractVector, y0 :: AbstractVector; kwargs...) where FC <: FloatOrComplex
-  solver = TrilqrSolver(A, b)
-  trilqr!(solver, A, b, c, x0, y0; kwargs...)
-  return (solver.x, solver.y, solver.stats)
-end
-
-function trilqr(A, b :: AbstractVector{FC}, c :: AbstractVector{FC}; kwargs...) where FC <: FloatOrComplex
-  solver = TrilqrSolver(A, b)
-  trilqr!(solver, A, b, c; kwargs...)
-  return (solver.x, solver.y, solver.stats)
-end
-
 """
     solver = trilqr!(solver::TrilqrSolver, A, b, c; kwargs...)
     solver = trilqr!(solver::TrilqrSolver, A, b, c, x0, y0; kwargs...)
@@ -91,11 +79,39 @@ See [`TrilqrSolver`](@ref) for more details about the `solver`.
 """
 function trilqr! end
 
-function trilqr!(solver :: TrilqrSolver{T,FC,S}, A, b :: AbstractVector{FC}, c :: AbstractVector{FC},
-                x0 :: AbstractVector, y0 :: AbstractVector; kwargs...) where {T <: AbstractFloat, FC <: FloatOrComplex{T}, S <: AbstractVector{FC}}
-  warm_start!(solver, x0, y0)
-  trilqr!(solver, A, b, c; kwargs...)
-  return solver
+def_kwargs_trilqr = (:(; transfer_to_usymcg::Bool = true),
+                     :(; atol::T = √eps(T)              ),
+                     :(; rtol::T = √eps(T)              ),
+                     :(; itmax::Int = 0                 ),
+                     :(; timemax::Float64 = Inf         ),
+                     :(; verbose::Int = 0               ),
+                     :(; history::Bool = false          ),
+                     :(; callback = solver -> false     ),
+                     :(; iostream::IO = kstdout         ))
+
+def_kwargs_trilqr = reduce(vcat, kw.args[1].args for kw in def_kwargs_trilqr)
+
+kwargs_trilqr = (:transfer_to_usymcg, :atol, :rtol, :itmax, :timemax, :verbose, :history, :callback, :iostream)
+
+@eval begin
+  function trilqr(A, b :: AbstractVector{FC}, c :: AbstractVector{FC}, x0 :: AbstractVector, y0 :: AbstractVector; $(def_kwargs_trilqr...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}}
+    solver = TrilqrSolver(A, b)
+    trilqr!(solver, A, b, c, x0, y0; $(kwargs_trilqr...))
+    return (solver.x, solver.y, solver.stats)
+  end
+
+  function trilqr(A, b :: AbstractVector{FC}, c :: AbstractVector{FC}; $(def_kwargs_trilqr...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}}
+    solver = TrilqrSolver(A, b)
+    trilqr!(solver, A, b, c; $(kwargs_trilqr...))
+    return (solver.x, solver.y, solver.stats)
+  end
+
+  function trilqr!(solver :: TrilqrSolver{T,FC,S}, A, b :: AbstractVector{FC}, c :: AbstractVector{FC},
[…]
