+++ src/cgls.jl
@@ -61,7 +61,7 @@ but simpler to implement.
 
 #### Keyword arguments
 
-* `N`: linear operator that models a Hermitian positive-definite matrix of size `n` used for preconditioning;
+* `M`: linear operator that models a Hermitian positive-definite matrix of size `n` used for preconditioning;
 * `ldiv`: define whether the preconditioner uses `ldiv!` or `mul!`;
 * `radius`: add the trust-region constraint ‖x‖ ≤ `radius` if `radius > 0`. Useful to compute a step in a trust-region method for optimization;
 * `λ`: regularization parameter;
@@ -86,12 +86,6 @@ but simpler to implement.
 """
 function cgls end
 
-function cgls(A, b :: AbstractVector{FC}; kwargs...) where FC <: FloatOrComplex
-  solver = CglsSolver(A, b)
-  cgls!(solver, A, b; kwargs...)
-  return (solver.x, solver.stats)
-end
-
 """
     solver = cgls!(solver::CglsSolver, A, b; kwargs...)
 
@@ -100,6 +94,31 @@ where `kwargs` are keyword arguments of [`cgls`](@ref).
 See [`CglsSolver`](@ref) for more details about the `solver`.
 """
 function cgls! end
+
+def_kwargs_cgls = (:(; M = I                     ),
+                   :(; ldiv::Bool = false        ),
+                   :(; radius::T = zero(T)       ),
+                   :(; λ::T = zero(T)            ),
+                   :(; atol::T = √eps(T)         ),
+                   :(; rtol::T = √eps(T)         ),
+                   :(; itmax::Int = 0            ),
+                   :(; timemax::Float64 = Inf    ),
+                   :(; verbose::Int = 0          ),
+                   :(; history::Bool = false     ),
+                   :(; callback = solver -> false),
+                   :(; iostream::IO = kstdout    ))
+
+def_kwargs_cgls = reduce(vcat, kw.args[1].args for kw in def_kwargs_cgls)
+
+kwargs_cgls = (:M, :ldiv, :radius, :λ, :atol, :rtol, :itmax, :timemax, :verbose, :history, :callback, :iostream)
+
+@eval begin
+  function cgls(A, b :: AbstractVector{FC}; $(def_kwargs_cgls...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}}
+    solver = CglsSolver(A, b)
+    cgls!(solver, A, b; $(kwargs_cgls...))
+    return (solver.x, solver.stats)
+  end
+end
 
 function cgls!(solver :: CglsSolver{T,FC,S}, A, b :: AbstractVector{FC};
                M=I, ldiv :: Bool=false, radius :: T=zero(T),
