### 2023-05-09 03695c4
Use metaprogramming to define the kwargs of all Krylov methods
