### 2023-05-10 5f546b6
Add an extract_parameters function
