+++ src/cgls.jl
@@ -108,7 +108,7 @@ def_kwargs_cgls = (:(; M = I                     ),
                    :(; callback = solver -> false),
                    :(; iostream::IO = kstdout    ))
 
-def_kwargs_cgls = reduce(vcat, kw.args[1].args for kw in def_kwargs_cgls)
+def_kwargs_cgls = mapreduce(extract_parameters, vcat, def_kwargs_cgls)
 
 kwargs_cgls = (:M, :ldiv, :radius, :λ, :atol, :rtol, :itmax, :timemax, :verbose, :history, :callback, :iostream)
 
