+++ src/cgne.jl
@@ -113,7 +113,7 @@ def_kwargs_cgne = (:(; N = I                     ),
                    :(; callback = solver -> false),
                    :(; iostream::IO = kstdout    ))
 
-def_kwargs_cgne = reduce(vcat, kw.args[1].args for kw in def_kwargs_cgne)
+def_kwargs_cgne = mapreduce(extract_parameters, vcat, def_kwargs_cgne)
 
 kwargs_cgne = (:N, :ldiv, :λ, :atol, :rtol, :itmax, :timemax, :verbose, :history, :callback, :iostream)
 
