+++ src/cgs.jl
@@ -102,7 +102,7 @@ def_kwargs_cgs = (:(; c::AbstractVector{FC} = b ),
                   :(; callback = solver -> false),
                   :(; iostream::IO = kstdout    ))
 
-def_kwargs_cgs = reduce(vcat, kw.args[1].args for kw in def_kwargs_cgs)
+def_kwargs_cgs = mapreduce(extract_parameters, vcat, def_kwargs_cgs)
 
 kwargs_cgs = (:c, :M, :N, :ldiv, :atol, :rtol, :itmax, :timemax, :verbose, :history, :callback, :iostream)
 
