### 2023-05-10 2f53b6e
Add an extract_parameters function
