+++ src/lsqr.jl
@@ -146,7 +146,7 @@ def_kwargs_lsqr = (:(; M = I                     ),
                    :(; callback = solver -> false),
                    :(; iostream::IO = kstdout    ))
 
-def_kwargs_lsqr = reduce(vcat, kw.args[1].args for kw in def_kwargs_lsqr)
+def_kwargs_lsqr = mapreduce(extract_parameters, vcat, def_kwargs_lsqr)
 
 kwargs_lsqr = (:M, :N, :ldiv, :sqd, :λ, :radius, :etol, :axtol, :btol, :conlim, :atol, :rtol, :itmax, :timemax, :verbose, :history, :callback, :iostream)
 
