+++ src/diom.jl
@@ -97,7 +97,7 @@ def_kwargs_diom = (:(; M = I                            ),
                    :(; callback = solver -> false       ),
                    :(; iostream::IO = kstdout           ))
 
-def_kwargs_diom = reduce(vcat, kw.args[1].args for kw in def_kwargs_diom)
+def_kwargs_diom = mapreduce(extract_parameters, vcat, def_kwargs_diom)
 
 kwargs_diom = (:M, :N, :ldiv, :reorthogonalization, :atol, :rtol, :itmax, :timemax, :verbose, :history, :callback, :iostream)
 
