+++ src/craigmr.jl
@@ -150,242 +150,239 @@ kwargs_craigmr = (:M, :N, :ldiv, :sqd, :λ, :atol, :rtol, :itmax, :timemax, :ver
     craigmr!(solver, A, b; $(kwargs_craigmr...))
     return (solver.x, solver.y, solver.stats)
   end
-end
 
-function craigmr!(solver :: CraigmrSolver{T,FC,S}, A, b :: AbstractVector{FC};
-                  M=I, N=I, ldiv :: Bool=false,
-                  sqd :: Bool=false, λ :: T=zero(T), atol :: T=√eps(T),
-                  rtol :: T=√eps(T), itmax :: Int=0,
-                  timemax :: Float64=Inf, verbose :: Int=0, history :: Bool=false,
-                  callback = solver -> false, iostream :: IO=kstdout) where {T <: AbstractFloat, FC <: FloatOrComplex{T}, S <: AbstractVector{FC}}
+  function craigmr!(solver :: CraigmrSolver{T,FC,S}, A, b :: AbstractVector{FC}; $(def_kwargs_craigmr...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}, S <: AbstractVector{FC}}
 
-  start_time = time_ns()
-  timemax_ns = 1e9 * timemax
-  m, n = size(A)
-  (m == solver.m && n == solver.n) || error("(solver.m, solver.n) = ($(solver.m), $(solver.n)) is inconsistent with size(A) = ($m, $n)")
-  length(b) == m || error("Inconsistent problem size")
-  (verbose > 0) && @printf(iostream, "CRAIGMR: system of %d equations in %d variables\n", m, n)
+    # Timer
+    start_time = time_ns()
+    timemax_ns = 1e9 * timemax
 
-  # Check sqd and λ parameters
-  sqd && (λ ≠ 0) && error("sqd cannot be set to true if λ ≠ 0 !")
-  sqd && (λ = one(T))
+    m, n = size(A)
+    (m == solver.m && n == solver.n) || error("(solver.m, solver.n) = ($(solver.m), $(solver.n)) is inconsistent with size(A) = ($m, $n)")
+    length(b) == m || error("Inconsistent problem size")
+    (verbose > 0) && @printf(iostream, "CRAIGMR: system of %d equations in %d variables\n", m, n)
 
-  # Tests M = Iₘ and N = Iₙ
-  MisI = (M === I)
-  NisI = (N === I)
+    # Check sqd and λ parameters
+    sqd && (λ ≠ 0) && error("sqd cannot be set to true if λ ≠ 0 !")
+    sqd && (λ = one(T))
 
-  # Check type consistency
-  eltype(A) == FC || error("eltype(A) ≠ $FC")
-  ktypeof(b) <: S || error("ktypeof(b) is not a subtype of $S")
+    # Tests M = Iₘ and N = Iₙ
+    MisI = (M === I)
+    NisI = (N === I)
 
-  # Compute the adjoint of A
-  Aᴴ = A'
+    # Check type consistency
+    eltype(A) == FC || error("eltype(A) ≠ $FC")
+    ktypeof(b) <: S || error("ktypeof(b) is not a subtype of $S")
 
-  # Set up workspace.
[…]
