+++ src/crls.jl
@@ -109,154 +109,152 @@ kwargs_crls = (:M, :ldiv, :radius, :λ, :atol, :rtol, :itmax, :timemax, :verbose
     crls!(solver, A, b; $(kwargs_crls...))
     return (solver.x, solver.stats)
   end
-end
 
-function crls!(solver :: CrlsSolver{T,FC,S}, A, b :: AbstractVector{FC};
-               M=I, ldiv :: Bool=false, radius :: T=zero(T),
-               λ :: T=zero(T), atol :: T=√eps(T), rtol :: T=√eps(T),
-               itmax :: Int=0, timemax :: Float64=Inf, verbose :: Int=0, history :: Bool=false,
-               callback = solver -> false, iostream :: IO=kstdout) where {T <: AbstractFloat, FC <: FloatOrComplex{T}, S <: AbstractVector{FC}}
+  function crls!(solver :: CrlsSolver{T,FC,S}, A, b :: AbstractVector{FC}; $(def_kwargs_crls...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}, S <: AbstractVector{FC}}
 
-  start_time = time_ns()
-  timemax_ns = 1e9 * timemax
-  m, n = size(A)
-  (m == solver.m && n == solver.n) || error("(solver.m, solver.n) = ($(solver.m), $(solver.n)) is inconsistent with size(A) = ($m, $n)")
-  length(b) == m || error("Inconsistent problem size")
-  (verbose > 0) && @printf(iostream, "CRLS: system of %d equations in %d variables\n", m, n)
+    # Timer
+    start_time = time_ns()
+    timemax_ns = 1e9 * timemax
 
-  # Tests M = Iₙ
-  MisI = (M === I)
+    m, n = size(A)
+    (m == solver.m && n == solver.n) || error("(solver.m, solver.n) = ($(solver.m), $(solver.n)) is inconsistent with size(A) = ($m, $n)")
+    length(b) == m || error("Inconsistent problem size")
+    (verbose > 0) && @printf(iostream, "CRLS: system of %d equations in %d variables\n", m, n)
 
-  # Check type consistency
-  eltype(A) == FC || error("eltype(A) ≠ $FC")
-  ktypeof(b) <: S || error("ktypeof(b) is not a subtype of $S")
+    # Tests M = Iₙ
+    MisI = (M === I)
 
-  # Compute the adjoint of A
-  Aᴴ = A'
+    # Check type consistency
+    eltype(A) == FC || error("eltype(A) ≠ $FC")
+    ktypeof(b) <: S || error("ktypeof(b) is not a subtype of $S")
 
-  # Set up workspace.
-  allocate_if(!MisI, solver, :Ms, S, m)
-  x, p, Ar, q = solver.x, solver.p, solver.Ar, solver.q
-  r, Ap, s, stats = solver.r, solver.Ap, solver.s, solver.stats
-  rNorms, ArNorms = stats.residuals, stats.Aresiduals
-  reset!(stats)
-  Ms  = MisI ? s  : solver.Ms
-  Mr  = MisI ? r  : solver.Ms
-  MAp = MisI ? Ap : solver.Ms
+    # Compute the adjoint of A
+    Aᴴ = A'
 
-  x .= zero(FC)
-  r .= b
-  bNorm = @knrm2(m, r)  # norm(b - A * x0) if x0 ≠ 0.
-  rNorm = bNorm  # + λ * ‖x0‖ if x0 ≠ 0 and λ > 0.
-  history && push!(rNorms, rNorm)
-  if bNorm == 0
-    stats.niter = 0
-    stats.solved, stats.inconsistent = true, false
[…]
