+++ src/cg.jl
@@ -115,156 +115,153 @@ kwargs_cg = (:M, :ldiv, :radius, :linesearch, :atol, :rtol, :itmax, :timemax, :v
     cg!(solver, A, b; $(kwargs_cg...))
     return solver
   end
-end
 
-function cg!(solver :: CgSolver{T,FC,S}, A, b :: AbstractVector{FC};
-             M=I, ldiv :: Bool=false, radius :: T=zero(T),
-             linesearch :: Bool=false, atol :: T=√eps(T),
-             rtol :: T=√eps(T), itmax :: Int=0,
-             timemax :: Float64=Inf, verbose :: Int=0, history :: Bool=false,
-             callback = solver -> false, iostream :: IO=kstdout) where {T <: AbstractFloat, FC <: FloatOrComplex{T}, S <: AbstractVector{FC}}
+  function cg!(solver :: CgSolver{T,FC,S}, A, b :: AbstractVector{FC}; $(def_kwargs_cg...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}, S <: AbstractVector{FC}}
 
-  start_time = time_ns()
-  timemax_ns = 1e9 * timemax
-  m, n = size(A)
-  (m == solver.m && n == solver.n) || error("(solver.m, solver.n) = ($(solver.m), $(solver.n)) is inconsistent with size(A) = ($m, $n)")
-  m == n || error("System must be square")
-  length(b) == n || error("Inconsistent problem size")
-  linesearch && (radius > 0) && error("`linesearch` set to `true` but trust-region radius > 0")
-  (verbose > 0) && @printf(iostream, "CG: system of %d equations in %d variables\n", n, n)
+    # Timer
+    start_time = time_ns()
+    timemax_ns = 1e9 * timemax
 
-  # Tests M = Iₙ
-  MisI = (M === I)
+    m, n = size(A)
+    (m == solver.m && n == solver.n) || error("(solver.m, solver.n) = ($(solver.m), $(solver.n)) is inconsistent with size(A) = ($m, $n)")
+    m == n || error("System must be square")
+    length(b) == n || error("Inconsistent problem size")
+    linesearch && (radius > 0) && error("`linesearch` set to `true` but trust-region radius > 0")
+    (verbose > 0) && @printf(iostream, "CG: system of %d equations in %d variables\n", n, n)
 
-  # Check type consistency
-  eltype(A) == FC || error("eltype(A) ≠ $FC")
-  ktypeof(b) <: S || error("ktypeof(b) is not a subtype of $S")
+    # Tests M = Iₙ
+    MisI = (M === I)
 
-  # Set up workspace.
-  allocate_if(!MisI, solver, :z, S, n)
-  Δx, x, r, p, Ap, stats = solver.Δx, solver.x, solver.r, solver.p, solver.Ap, solver.stats
-  warm_start = solver.warm_start
-  rNorms = stats.residuals
-  reset!(stats)
-  z = MisI ? r : solver.z
+    # Check type consistency
+    eltype(A) == FC || error("eltype(A) ≠ $FC")
+    ktypeof(b) <: S || error("ktypeof(b) is not a subtype of $S")
 
-  x .= zero(FC)
-  if warm_start
-    mul!(r, A, Δx)
-    @kaxpby!(n, one(FC), b, -one(FC), r)
-  else
-    r .= b
-  end
-  MisI || mulorldiv!(z, M, r, ldiv)
-  p .= z
-  γ = @kdotr(n, r, z)
-  rNorm = sqrt(γ)
-  history && push!(rNorms, rNorm)
-  if γ == 0
-    stats.niter = 0
-    stats.solved, stats.inconsistent = true, false
[…]
