+++ src/diom.jl
@@ -119,212 +119,209 @@ kwargs_diom = (:M, :N, :ldiv, :reorthogonalization, :atol, :rtol, :itmax, :timem
     diom!(solver, A, b; $(kwargs_diom...))
     return solver
   end
-end
 
-function diom!(solver :: DiomSolver{T,FC,S}, A, b :: AbstractVector{FC};
-               M=I, N=I, ldiv :: Bool=false,
-               reorthogonalization :: Bool=false, atol :: T=√eps(T),
-               rtol :: T=√eps(T), itmax :: Int=0,
-               timemax :: Float64=Inf, verbose :: Int=0, history :: Bool=false,
-               callback = solver -> false, iostream :: IO=kstdout) where {T <: AbstractFloat, FC <: FloatOrComplex{T}, S <: AbstractVector{FC}}
+  function diom!(solver :: DiomSolver{T,FC,S}, A, b :: AbstractVector{FC}; $(def_kwargs_diom...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}, S <: AbstractVector{FC}}
 
-  start_time = time_ns()
-  timemax_ns = 1e9 * timemax
-  m, n = size(A)
-  (m == solver.m && n == solver.n) || error("(solver.m, solver.n) = ($(solver.m), $(solver.n)) is inconsistent with size(A) = ($m, $n)")
-  m == n || error("System must be square")
-  length(b) == m || error("Inconsistent problem size")
-  (verbose > 0) && @printf(iostream, "DIOM: system of size %d\n", n)
+    # Timer
+    start_time = time_ns()
+    timemax_ns = 1e9 * timemax
 
-  # Check M = Iₙ and N = Iₙ
-  MisI = (M === I)
-  NisI = (N === I)
+    m, n = size(A)
+    (m == solver.m && n == solver.n) || error("(solver.m, solver.n) = ($(solver.m), $(solver.n)) is inconsistent with size(A) = ($m, $n)")
+    m == n || error("System must be square")
+    length(b) == m || error("Inconsistent problem size")
+    (verbose > 0) && @printf(iostream, "DIOM: system of size %d\n", n)
 
-  # Check type consistency
-  eltype(A) == FC || error("eltype(A) ≠ $FC")
-  ktypeof(b) <: S || error("ktypeof(b) is not a subtype of $S")
+    # Check M = Iₙ and N = Iₙ
+    MisI = (M === I)
+    NisI = (N === I)
 
-  # Set up workspace.
-  allocate_if(!MisI, solver, :w, S, n)
-  allocate_if(!NisI, solver, :z, S, n)
-  Δx, x, t, P, V = solver.Δx, solver.x, solver.t, solver.P, solver.V
-  L, H, stats = solver.L, solver.H, solver.stats
-  warm_start = solver.warm_start
-  rNorms = stats.residuals
-  reset!(stats)
-  w  = MisI ? t : solver.w
-  r₀ = MisI ? t : solver.w
+    # Check type consistency
+    eltype(A) == FC || error("eltype(A) ≠ $FC")
+    ktypeof(b) <: S || error("ktypeof(b) is not a subtype of $S")
 
-  # Initial solution x₀ and residual r₀.
-  x .= zero(FC)  # x₀
-  if warm_start
-    mul!(t, A, Δx)
[…]
