### 2023-05-10 4821fca
Display a warning when the operator and the right-hand side don't have the same type
