### 2023-05-10 b901092
Display a warning when the operator and the right-hand side don't have the same type
