### 2023-05-10 4363bd6
Combine the allocation of the KrylovSolver with the warm-start
