### 2023-05-10 3d3fe56
Take into account the time to allocate the KrylovSolver
and perform warm-start when the timemax option is used
