+++ src/dqgmres.jl
@@ -103,20 +103,26 @@ kwargs_dqgmres = (:M, :N, :ldiv, :reorthogonalization, :atol, :rtol, :itmax, :ti
 
 @eval begin
   function dqgmres(A, b :: AbstractVector{FC}, x0 :: AbstractVector; memory :: Int=20, $(def_kwargs_dqgmres...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}}
+    start_time = time_ns()
     solver = DqgmresSolver(A, b, memory)
     warm_start!(solver, x0)
+    timemax -= (time_ns() - start_time) / 1e9
     dqgmres!(solver, A, b; $(kwargs_dqgmres...))
     return (solver.x, solver.stats)
   end
 
   function dqgmres(A, b :: AbstractVector{FC}; memory :: Int=20, $(def_kwargs_dqgmres...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}}
+    start_time = time_ns()
     solver = DqgmresSolver(A, b, memory)
+    timemax -= (time_ns() - start_time) / 1e9
     dqgmres!(solver, A, b; $(kwargs_dqgmres...))
     return (solver.x, solver.stats)
   end
 
   function dqgmres!(solver :: DqgmresSolver{T,FC,S}, A, b :: AbstractVector{FC}, x0 :: AbstractVector; $(def_kwargs_dqgmres...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}, S <: AbstractVector{FC}}
+    start_time = time_ns()
     warm_start!(solver, x0)
+    timemax -= (time_ns() - start_time) / 1e9
     dqgmres!(solver, A, b; $(kwargs_dqgmres...))
     return solver
   end
