+++ src/craig.jl
@@ -157,7 +157,9 @@ kwargs_craig = (:M, :N, :ldiv, :transfer_to_lsqr, :sqd, :λ, :btol, :conlim, :at
 
 @eval begin
   function craig(A, b :: AbstractVector{FC}; $(def_kwargs_craig...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}}
+    start_time = time_ns()
     solver = CraigSolver(A, b)
+    timemax -= (time_ns() - start_time) / 1e9
     craig!(solver, A, b; $(kwargs_craig...))
     return (solver.x, solver.y, solver.stats)
   end
