+++ src/bilq.jl
@@ -92,20 +92,26 @@ kwargs_bilq = (:c, :transfer_to_bicg, :atol, :rtol, :itmax, :timemax, :verbose, 
 
 @eval begin
   function bilq(A, b :: AbstractVector{FC}, x0 :: AbstractVector; $(def_kwargs_bilq...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}}
+    start_time = time_ns()
     solver = BilqSolver(A, b)
     warm_start!(solver, x0)
+    timemax -= (time_ns() - start_time) / 1e9
     bilq!(solver, A, b; $(kwargs_bilq...))
     return (solver.x, solver.stats)
   end
 
   function bilq(A, b :: AbstractVector{FC}; $(def_kwargs_bilq...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}}
+    start_time = time_ns()
     solver = BilqSolver(A, b)
+    timemax -= (time_ns() - start_time) / 1e9
     bilq!(solver, A, b; $(kwargs_bilq...))
     return (solver.x, solver.stats)
   end
 
   function bilq!(solver :: BilqSolver{T,FC,S}, A, b :: AbstractVector{FC}, x0 :: AbstractVector; $(def_kwargs_bilq...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}, S <: AbstractVector{FC}}
+    start_time = time_ns()
     warm_start!(solver, x0)
+    timemax -= (time_ns() - start_time) / 1e9
     bilq!(solver, A, b; $(kwargs_bilq...))
     return solver
   end
