### 2023-05-10 5f7ea0f
Take into account the time to allocate the KrylovSolver
and perform warm-start when the timemax option is used
