+++ src/gmres.jl
@@ -99,20 +99,26 @@ kwargs_gmres = (:M, :N, :ldiv, :restart, :reorthogonalization, :atol, :rtol, :it
 
 @eval begin
   function gmres(A, b :: AbstractVector{FC}, x0 :: AbstractVector; memory :: Int=20, $(def_kwargs_gmres...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}}
+    start_time = time_ns()
     solver = GmresSolver(A, b, memory)
     warm_start!(solver, x0)
+    timemax -= (time_ns() - start_time) / 1e9
     gmres!(solver, A, b; $(kwargs_gmres...))
     return (solver.x, solver.stats)
   end
 
   function gmres(A, b :: AbstractVector{FC}; memory :: Int=20, $(def_kwargs_gmres...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}}
+    start_time = time_ns()
     solver = GmresSolver(A, b, memory)
+    timemax -= (time_ns() - start_time) / 1e9
     gmres!(solver, A, b; $(kwargs_gmres...))
     return (solver.x, solver.stats)
   end
 
   function gmres!(solver :: GmresSolver{T,FC,S}, A, b :: AbstractVector{FC}, x0 :: AbstractVector; $(def_kwargs_gmres...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}, S <: AbstractVector{FC}}
+    start_time = time_ns()
     warm_start!(solver, x0)
+    timemax -= (time_ns() - start_time) / 1e9
     gmres!(solver, A, b; $(kwargs_gmres...))
     return solver
   end
