+++ src/cr.jl
@@ -106,20 +106,26 @@ kwargs_cr = (:M, :ldiv, :radius, :linesearch, :γ, :atol, :rtol, :itmax, :timema
 
 @eval begin
   function cr(A, b :: AbstractVector{FC}, x0 :: AbstractVector; $(def_kwargs_cr...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}}
+    start_time = time_ns()
     solver = CrSolver(A, b)
     warm_start!(solver, x0)
+    timemax -= (time_ns() - start_time) / 1e9
     cr!(solver, A, b; $(kwargs_cr...))
     return (solver.x, solver.stats)
   end
 
   function cr(A, b :: AbstractVector{FC}; $(def_kwargs_cr...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}}
+    start_time = time_ns()
     solver = CrSolver(A, b)
+    timemax -= (time_ns() - start_time) / 1e9
     cr!(solver, A, b; $(kwargs_cr...))
     return (solver.x, solver.stats)
   end
 
   function cr!(solver :: CrSolver{T,FC,S}, A, b :: AbstractVector{FC}, x0 :: AbstractVector; $(def_kwargs_cr...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}, S <: AbstractVector{FC}}
+    start_time = time_ns()
     warm_start!(solver, x0)
+    timemax -= (time_ns() - start_time) / 1e9
     cr!(solver, A, b; $(kwargs_cr...))
     return solver
   end
