### 2023-05-11 ba849b4
Test the verbose mode
