+++ test/runtests.jl
@@ -44,3 +44,4 @@ include("test_allocations.jl")
 include("test_mp.jl")
 include("test_solvers.jl")
 include("test_warm_start.jl")
+include("test_verbose.jl")
