### 2023-05-11 113f325
Test the verbose mode
