### 2023-05-12 993f1c0
Update the verbose mode to display a timer
