### 2023-05-12 b62f606
Update the verbose mode to display a timer
