### 2023-05-12 55f577c
Add a field timer in the statistics
