+++ test/test_stats.jl
@@ -1,5 +1,5 @@
 @testset "stats" begin
-  stats = Krylov.SimpleStats(0, true, true, Float64[1.0], Float64[2.0], Float64[], "unknown")
+  stats = Krylov.SimpleStats(0, true, true, Float64[1.0], Float64[2.0], Float64[], 1.234, "unknown")
   io = IOBuffer()
   show(io, stats)
   showed = String(take!(io))
@@ -11,6 +11,7 @@
   residuals: [ 1.0e+00 ]
   Aresiduals: [ 2.0e+00 ]
   κ₂(A): []
+  timer: 1.23s
   status: unknown"""
   @test strip.(split(chomp(showed), "\n")) == strip.(split(chomp(expected), "\n"))
   Krylov.reset!(stats)
@@ -18,7 +19,7 @@
   nbytes_allocated = @allocated Krylov.reset!(stats)
   @test nbytes_allocated == 0
 
-  stats = Krylov.LsmrStats(0, true, true, Float64[1.0], Float64[2.0], Float64(3.0), Float64(4.0), Float64(5.0), Float64(6.0), Float64(7.0), "unknown")
+  stats = Krylov.LsmrStats(0, true, true, Float64[1.0], Float64[2.0], Float64(3.0), Float64(4.0), Float64(5.0), Float64(6.0), Float64(7.0), 0.1234, "unknown")
   io = IOBuffer()
   show(io, stats)
   showed = String(take!(io))
@@ -34,6 +35,7 @@
   κ₂(A): 5.0
   ‖A‖F: 6.0
   xNorm: 7.0
+  timer: 123.40ms
   status: unknown"""
   @test strip.(split(chomp(showed), "\n")) == strip.(split(chomp(expected), "\n"))
   Krylov.reset!(stats)
@@ -41,7 +43,7 @@
   nbytes_allocated = @allocated Krylov.reset!(stats)
   @test nbytes_allocated == 0
 
-  stats = Krylov.LanczosStats(0, true, Float64[3.0], true, NaN, NaN, "unknown")
+  stats = Krylov.LanczosStats(0, true, Float64[3.0], true, NaN, NaN, 1.234, "unknown")
   io = IOBuffer()
   show(io, stats)
   showed = String(take!(io))
@@ -53,6 +55,7 @@
   indefinite: true
   ‖A‖F: NaN
   κ₂(A): NaN
+  timer: 1.23s
   status: unknown"""
   @test strip.(split(chomp(showed), "\n")) == strip.(split(chomp(expected), "\n"))
   Krylov.reset!(stats)
@@ -60,7 +63,7 @@
   nbytes_allocated = @allocated Krylov.reset!(stats)
   @test nbytes_allocated == 0
 
-  stats = Krylov.LanczosShiftStats(0, true, [Float64[0.9, 0.5], Float64[0.6, 0.4, 0.1]], BitVector([false, true]), NaN, NaN, "unknown")
+  stats = Krylov.LanczosShiftStats(0, true, [Float64[0.9, 0.5], Float64[0.6, 0.4, 0.1]], BitVector([false, true]), NaN, NaN, 0.00056789, "unknown")
   io = IOBuffer()
   show(io, stats)
   showed = String(take!(io))
@@ -72,13 +75,14 @@
   indefinite: Bool[0, 1]
   ‖A‖F: NaN
   κ₂(A): NaN
+  timer: 567.89μs
   status: unknown"""
   @test strip.(split(chomp(showed), "\n")) == strip.(split(chomp(expected), "\n"))
   Krylov.reset!(stats)
   nbytes_allocated = @allocated Krylov.reset!(stats)
   @test nbytes_allocated == 0
 
[…]
