+++ src/krylov_solvers.jl
@@ -87,7 +87,7 @@ function MinresSolver(m, n, S; window :: Int=5)
   y  = S(undef, n)
   v  = S(undef, 0)
   err_vec = zeros(T, window)
-  stats = SimpleStats(0, false, false, T[], T[], T[], "unknown")
+  stats = SimpleStats(0, false, false, T[], T[], T[], 0.0, "unknown")
   solver = MinresSolver{T,FC,S}(m, n, Δx, x, r1, r2, w1, w2, y, v, err_vec, false, stats)
   return solver
 end
@@ -130,7 +130,7 @@ function CgSolver(m, n, S)
   p  = S(undef, n)
   Ap = S(undef, n)
   z  = S(undef, 0)
-  stats = SimpleStats(0, false, false, T[], T[], T[], "unknown")
+  stats = SimpleStats(0, false, false, T[], T[], T[], 0.0, "unknown")
   solver = CgSolver{T,FC,S}(m, n, Δx, x, r, p, Ap, z, false, stats)
   return solver
 end
@@ -175,7 +175,7 @@ function CrSolver(m, n, S)
   q  = S(undef, n)
   Ar = S(undef, n)
   Mq = S(undef, 0)
-  stats = SimpleStats(0, false, false, T[], T[], T[], "unknown")
+  stats = SimpleStats(0, false, false, T[], T[], T[], 0.0, "unknown")
   solver = CrSolver{T,FC,S}(m, n, Δx, x, r, p, q, Ar, Mq, false, stats)
   return solver
 end
@@ -226,7 +226,7 @@ function SymmlqSolver(m, n, S; window :: Int=5)
   clist   = zeros(T, window)
   zlist   = zeros(T, window)
   sprod   = ones(T, window)
-  stats = SymmlqStats(0, false, T[], Union{T, Missing}[], T[], Union{T, Missing}[], T(NaN), T(NaN), "unknown")
+  stats = SymmlqStats(0, false, T[], Union{T, Missing}[], T[], Union{T, Missing}[], T(NaN), T(NaN), 0.0, "unknown")
   solver = SymmlqSolver{T,FC,S}(m, n, Δx, x, Mvold, Mv, Mv_next, w̅, v, clist, zlist, sprod, false, stats)
   return solver
 end
@@ -271,7 +271,7 @@ function CgLanczosSolver(m, n, S)
   p       = S(undef, n)
   Mv_next = S(undef, n)
   v       = S(undef, 0)
-  stats = LanczosStats(0, false, T[], false, T(NaN), T(NaN), "unknown")
+  stats = LanczosStats(0, false, T[], false, T(NaN), T(NaN), 0.0, "unknown")
   solver = CgLanczosSolver{T,FC,S}(m, n, Δx, x, Mv, Mv_prev, p, Mv_next, v, false, stats)
   return solver
 end
@@ -329,7 +329,7 @@ function CgLanczosShiftSolver(m, n, nshifts, S)
   indefinite = BitVector(undef, nshifts)
   converged  = BitVector(undef, nshifts)
   not_cv     = BitVector(undef, nshifts)
-  stats = LanczosShiftStats(0, false, Vector{T}[T[] for i = 1 : nshifts], indefinite, T(NaN), T(NaN), "unknown")
+  stats = LanczosShiftStats(0, false, Vector{T}[T[] for i = 1 : nshifts], indefinite, T(NaN), T(NaN), 0.0, "unknown")
[…]
