### 2023-05-12 abcbe23
Update stats.timer in all Krylov methods
