+++ src/lnlq.jl
@@ -211,6 +211,7 @@ kwargs_lnlq = (:M, :N, :ldiv, :transfer_to_craig, :sqd, :λ, :σ, :utolx, :utoly
       stats.solved = true
       stats.error_with_bnd = false
       history && push!(rNorms, bNorm)
+      stats.timer = ktimer(start_time)
       stats.status = "x = 0 is a zero-residual solution"
       return solver
     end
@@ -546,6 +547,7 @@ kwargs_lnlq = (:M, :N, :ldiv, :transfer_to_craig, :sqd, :λ, :σ, :utolx, :utoly
     stats.niter = iter
     stats.solved = solved_lq || solved_cg
     stats.error_with_bnd = complex_error_bnd
+    stats.timer = ktimer(start_time)
     stats.status = status
     return solver
   end
