### 2023-05-12 fa0da9c
Update stats.timer in all Krylov methods
