+++ src/trimr.jl
@@ -567,6 +567,7 @@ kwargs_trimr = (:M, :N, :ldiv, :spd, :snd, :flip, :sp, :τ, :ν, :atol, :rtol, :
     stats.niter = iter
     stats.solved = solved
     stats.inconsistent = !solved && breakdown
+    stats.timer = ktimer(start_time)
     stats.status = status
     return solver
   end
