### 2023-05-15 6d9def6
Improve the reliability of solver.stats.timer
