+++ src/cr.jl
@@ -109,24 +109,30 @@ kwargs_cr = (:M, :ldiv, :radius, :linesearch, :γ, :atol, :rtol, :itmax, :timema
     start_time = time_ns()
     solver = CrSolver(A, b)
     warm_start!(solver, x0)
-    timemax -= ktimer(start_time)
+    elapsed_time = ktimer(start_time)
+    timemax -= elapsed_time
     cr!(solver, A, b; $(kwargs_cr...))
+    solver.stats.timer += elapsed_time
     return (solver.x, solver.stats)
   end
 
   function cr(A, b :: AbstractVector{FC}; $(def_kwargs_cr...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}}
     start_time = time_ns()
     solver = CrSolver(A, b)
-    timemax -= ktimer(start_time)
+    elapsed_time = ktimer(start_time)
+    timemax -= elapsed_time
     cr!(solver, A, b; $(kwargs_cr...))
+    solver.stats.timer += elapsed_time
     return (solver.x, solver.stats)
   end
 
   function cr!(solver :: CrSolver{T,FC,S}, A, b :: AbstractVector{FC}, x0 :: AbstractVector; $(def_kwargs_cr...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}, S <: AbstractVector{FC}}
     start_time = time_ns()
     warm_start!(solver, x0)
-    timemax -= ktimer(start_time)
+    elapsed_time = ktimer(start_time)
+    timemax -= elapsed_time
     cr!(solver, A, b; $(kwargs_cr...))
+    solver.stats.timer += elapsed_time
     return solver
   end
 
