+++ src/craigmr.jl
@@ -148,8 +148,10 @@ kwargs_craigmr = (:M, :N, :ldiv, :sqd, :λ, :atol, :rtol, :itmax, :timemax, :ver
   function craigmr(A, b :: AbstractVector{FC}; $(def_kwargs_craigmr...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}}
     start_time = time_ns()
     solver = CraigmrSolver(A, b)
-    timemax -= ktimer(start_time)
+    elapsed_time = ktimer(start_time)
+    timemax -= elapsed_time
     craigmr!(solver, A, b; $(kwargs_craigmr...))
+    solver.stats.timer += elapsed_time
     return (solver.x, solver.y, solver.stats)
   end
 
