+++ src/tricg.jl
@@ -129,24 +129,30 @@ kwargs_tricg = (:M, :N, :ldiv, :spd, :snd, :flip, :τ, :ν, :atol, :rtol, :itmax
     start_time = time_ns()
     solver = TricgSolver(A, b)
     warm_start!(solver, x0, y0)
-    timemax -= ktimer(start_time)
+    elapsed_time = ktimer(start_time)
+    timemax -= elapsed_time
     tricg!(solver, A, b, c; $(kwargs_tricg...))
+    solver.stats.timer += elapsed_time
     return (solver.x, solver.y, solver.stats)
   end
 
   function tricg(A, b :: AbstractVector{FC}, c :: AbstractVector{FC}; $(def_kwargs_tricg...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}}
     start_time = time_ns()
     solver = TricgSolver(A, b)
-    timemax -= ktimer(start_time)
+    elapsed_time = ktimer(start_time)
+    timemax -= elapsed_time
     tricg!(solver, A, b, c; $(kwargs_tricg...))
+    solver.stats.timer += elapsed_time
     return (solver.x, solver.y, solver.stats)
   end
 
   function tricg!(solver :: TricgSolver{T,FC,S}, A, b :: AbstractVector{FC}, c :: AbstractVector{FC}, x0 :: AbstractVector, y0 :: AbstractVector; $(def_kwargs_tricg...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}, S <: AbstractVector{FC}}
     start_time = time_ns()
     warm_start!(solver, x0, y0)
-    timemax -= ktimer(start_time)
+    elapsed_time = ktimer(start_time)
+    timemax -= elapsed_time
     tricg!(solver, A, b, c; $(kwargs_tricg...))
+    solver.stats.timer += elapsed_time
     return solver
   end
 
