+++ src/bicgstab.jl
@@ -110,24 +110,30 @@ kwargs_bicgstab = (:c, :M, :N, :ldiv, :atol, :rtol, :itmax, :timemax, :verbose, 
     start_time = time_ns()
     solver = BicgstabSolver(A, b)
     warm_start!(solver, x0)
-    timemax -= ktimer(start_time)
+    elapsed_time = ktimer(start_time)
+    timemax -= elapsed_time
     bicgstab!(solver, A, b; $(kwargs_bicgstab...))
+    solver.stats.timer += elapsed_time
     return (solver.x, solver.stats)
   end
 
   function bicgstab(A, b :: AbstractVector{FC}; $(def_kwargs_bicgstab...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}}
     start_time = time_ns()
     solver = BicgstabSolver(A, b)
-    timemax -= ktimer(start_time)
+    elapsed_time = ktimer(start_time)
+    timemax -= elapsed_time
     bicgstab!(solver, A, b; $(kwargs_bicgstab...))
+    solver.stats.timer += elapsed_time
     return (solver.x, solver.stats)
   end
 
   function bicgstab!(solver :: BicgstabSolver{T,FC,S}, A, b :: AbstractVector{FC}, x0 :: AbstractVector; $(def_kwargs_bicgstab...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}, S <: AbstractVector{FC}}
     start_time = time_ns()
     warm_start!(solver, x0)
-    timemax -= ktimer(start_time)
+    elapsed_time = ktimer(start_time)
+    timemax -= elapsed_time
     bicgstab!(solver, A, b; $(kwargs_bicgstab...))
+    solver.stats.timer += elapsed_time
     return solver
   end
 

[…]
