### 2023-05-15 b9d1f0a
Improve the reliability of solver.stats.timer
