+++ src/fgmres.jl
@@ -109,24 +109,30 @@ kwargs_fgmres = (:M, :N, :ldiv, :restart, :reorthogonalization, :atol, :rtol, :i
     start_time = time_ns()
     solver = FgmresSolver(A, b, memory)
     warm_start!(solver, x0)
-    timemax -= ktimer(start_time)
+    elapsed_time = ktimer(start_time)
+    timemax -= elapsed_time
     fgmres!(solver, A, b; $(kwargs_fgmres...))
+    solver.stats.timer += elapsed_time
     return (solver.x, solver.stats)
   end
 
   function fgmres(A, b :: AbstractVector{FC}; memory :: Int=20, $(def_kwargs_fgmres...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}}
     start_time = time_ns()
     solver = FgmresSolver(A, b, memory)
-    timemax -= ktimer(start_time)
+    elapsed_time = ktimer(start_time)
+    timemax -= elapsed_time
     fgmres!(solver, A, b; $(kwargs_fgmres...))
+    solver.stats.timer += elapsed_time
     return (solver.x, solver.stats)
   end
 
   function fgmres!(solver :: FgmresSolver{T,FC,S}, A, b :: AbstractVector{FC}, x0 :: AbstractVector; $(def_kwargs_fgmres...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}, S <: AbstractVector{FC}}
     start_time = time_ns()
     warm_start!(solver, x0)
-    timemax -= ktimer(start_time)
+    elapsed_time = ktimer(start_time)
+    timemax -= elapsed_time
     fgmres!(solver, A, b; $(kwargs_fgmres...))
+    solver.stats.timer += elapsed_time
     return solver
   end
 
