### 2023-05-17 18eec51
Keep the old behaviour of ktypeof for FillVector
