### 2023-06-28 4723b13
Optimize the generic function solve!
