+++ src/diom.jl
@@ -84,6 +84,9 @@ See [`DiomSolver`](@ref) for more details about the `solver`.
 """
 function diom! end
 
+def_args_diom = (:(A                    ),
+                 :(b::AbstractVector{FC}))
+
 def_kwargs_diom = (:(; M = I                            ),
                    :(; N = I                            ),
                    :(; ldiv::Bool = false               ),
@@ -99,6 +102,7 @@ def_kwargs_diom = (:(; M = I                            ),
 
 def_kwargs_diom = mapreduce(extract_parameters, vcat, def_kwargs_diom)
 
+args_diom = (:A, :b)
 kwargs_diom = (:M, :N, :ldiv, :reorthogonalization, :atol, :rtol, :itmax, :timemax, :verbose, :history, :callback, :iostream)
 
 @eval begin
