+++ src/cg.jl
@@ -80,6 +80,9 @@ See [`CgSolver`](@ref) for more details about the `solver`.
 """
 function cg! end
 
+def_args_cg = (:(A                    ),
+               :(b::AbstractVector{FC}))
+
 def_kwargs_cg = (:(; M = I                     ),
                  :(; ldiv::Bool = false        ),
                  :(; radius::T = zero(T)       ),
@@ -95,6 +98,7 @@ def_kwargs_cg = (:(; M = I                     ),
 
 def_kwargs_cg = mapreduce(extract_parameters, vcat, def_kwargs_cg)
 
+args_cg = (:A, :b)
 kwargs_cg = (:M, :ldiv, :radius, :linesearch, :atol, :rtol, :itmax, :timemax, :verbose, :history, :callback, :iostream)
 
 @eval begin
