+++ src/usymqr.jl
@@ -87,6 +87,10 @@ See [`UsymqrSolver`](@ref) for more details about the `solver`.
 """
 function usymqr! end
 
+def_args_usymqr = (:(A                    ),
+                   :(b::AbstractVector{FC}),
+                   :(c::AbstractVector{FC}))
+
 def_kwargs_usymqr = (:(; atol::T = √eps(T)         ),
                      :(; rtol::T = √eps(T)         ),
                      :(; itmax::Int = 0            ),
@@ -98,6 +102,7 @@ def_kwargs_usymqr = (:(; atol::T = √eps(T)         ),
 
 def_kwargs_usymqr = mapreduce(extract_parameters, vcat, def_kwargs_usymqr)
 
+args_usymqr = (:A, :b, :c)
 kwargs_usymqr = (:atol, :rtol, :itmax, :timemax, :verbose, :history, :callback, :iostream)
 
 @eval begin

[…]
