### 2023-06-28 1a1c59f
Optimize the generic function solve!
