+++ src/minres_qlp.jl
@@ -84,6 +84,9 @@ See [`MinresQlpSolver`](@ref) for more details about the `solver`.
 """
 function minres_qlp! end
 
+def_args_minres_qlp = (:(A                    ),
+                       :(b::AbstractVector{FC}))
+
 def_kwargs_minres_qlp = (:(; M = I                     ),
                          :(; ldiv::Bool = false        ),
                          :(; λ::T = zero(T)            ),
@@ -99,6 +102,7 @@ def_kwargs_minres_qlp = (:(; M = I                     ),
 
 def_kwargs_minres_qlp = mapreduce(extract_parameters, vcat, def_kwargs_minres_qlp)
 
+args_minres_qlp = (:A, :b)
 kwargs_minres_qlp = (:M, :ldiv, :λ, :atol, :rtol, :Artol, :itmax, :timemax, :verbose, :history, :callback, :iostream)
 
 @eval begin
