### 2023-06-28 0de0116
[CI] Remove the build with Linux ARM
