### 2023-06-28 dca42a5
Add methods solve! with optional arguments
