+++ src/qmr.jl
@@ -86,6 +86,8 @@ function qmr! end
 def_args_qmr = (:(A                    ),
                 :(b::AbstractVector{FC}))
 
+def_optargs_qmr = (:(x0::AbstractVector),)
+
 def_kwargs_qmr = (:(; c::AbstractVector{FC} = b ),
                   :(; atol::T = √eps(T)         ),
                   :(; rtol::T = √eps(T)         ),
@@ -99,6 +101,7 @@ def_kwargs_qmr = (:(; c::AbstractVector{FC} = b ),
 def_kwargs_qmr = mapreduce(extract_parameters, vcat, def_kwargs_qmr)
 
 args_qmr = (:A, :b)
+optargs_qmr = (:x0,)
 kwargs_qmr = (:c, :atol, :rtol, :itmax, :timemax, :verbose, :history, :callback, :iostream)
 
 @eval begin
