+++ src/gmres.jl
@@ -82,6 +82,8 @@ function gmres! end
 def_args_gmres = (:(A                    ),
                   :(b::AbstractVector{FC}))
 
+def_optargs_gmres = (:(x0::AbstractVector),)
+
 def_kwargs_gmres = (:(; M = I                            ),
                     :(; N = I                            ),
                     :(; ldiv::Bool = false               ),
@@ -99,6 +101,7 @@ def_kwargs_gmres = (:(; M = I                            ),
 def_kwargs_gmres = mapreduce(extract_parameters, vcat, def_kwargs_gmres)
 
 args_gmres = (:A, :b)
+optargs_gmres = (:x0,)
 kwargs_gmres = (:M, :N, :ldiv, :restart, :reorthogonalization, :atol, :rtol, :itmax, :timemax, :verbose, :history, :callback, :iostream)
 
 @eval begin
