### 2023-06-28 e66349d
Add methods solve! with optional arguments
