+++ src/tricg.jl
@@ -107,6 +107,9 @@ def_args_tricg = (:(A                    ),
                   :(b::AbstractVector{FC}),
                   :(c::AbstractVector{FC}))
 
+def_optargs_tricg = (:(x0::AbstractVector),
+                     :(y0::AbstractVector))
+
 def_kwargs_tricg = (:(; M = I                     ),
                     :(; N = I                     ),
                     :(; ldiv::Bool = false        ),
@@ -127,6 +130,7 @@ def_kwargs_tricg = (:(; M = I                     ),
 def_kwargs_tricg = mapreduce(extract_parameters, vcat, def_kwargs_tricg)
 
 args_tricg = (:A, :b, :c)
+optargs_tricg = (:x0, :y0)
 kwargs_tricg = (:M, :N, :ldiv, :spd, :snd, :flip, :τ, :ν, :atol, :rtol, :itmax, :timemax, :verbose, :history, :callback, :iostream)
 
 @eval begin
