+++ src/cgs.jl
@@ -92,6 +92,8 @@ function cgs! end
 def_args_cgs = (:(A                    ),
                 :(b::AbstractVector{FC}))
 
+def_optargs_cgs = (:(x0::AbstractVector),)
+
 def_kwargs_cgs = (:(; c::AbstractVector{FC} = b ),
                   :(; M = I                     ),
                   :(; N = I                     ),
@@ -108,6 +110,7 @@ def_kwargs_cgs = (:(; c::AbstractVector{FC} = b ),
 def_kwargs_cgs = mapreduce(extract_parameters, vcat, def_kwargs_cgs)
 
 args_cgs = (:A, :b)
+optargs_cgs = (:x0,)
 kwargs_cgs = (:c, :M, :N, :ldiv, :atol, :rtol, :itmax, :timemax, :verbose, :history, :callback, :iostream)
 
 @eval begin
