### 2023-06-28 b3804c9
Add tests for in-place methods with keyword arguments
