### 2023-06-30 #5187 @dpo
@danielskatz @jedbrown There has been no activity here since April 20th. We submitted this paper on [November 26th, 2022](https://github.com/openjournals/joss-reviews/issues/4970), and it is now June 30th. More than seven months have gone by. I am fully aware that everyone is overcommitted, but is that really the kind of time frame we can expect when submitting to JOSS? I feel this submission has fallen into oblivion, and as an author, that does not feel good.

Thank you for your attention.
