### 2023-07-16 69e446f
Use metaprogramming to reduce the code
