+++ src/lsqr.jl
@@ -155,17 +155,17 @@ args_lsqr = (:A, :b)
 kwargs_lsqr = (:M, :N, :ldiv, :sqd, :λ, :radius, :etol, :axtol, :btol, :conlim, :atol, :rtol, :itmax, :timemax, :verbose, :history, :callback, :iostream)
 
 @eval begin
-  function lsqr(A, b :: AbstractVector{FC}; window :: Int=5, $(def_kwargs_lsqr...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}}
+  function lsqr($(def_args_lsqr...); window :: Int=5, $(def_kwargs_lsqr...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}}
     start_time = time_ns()
     solver = LsqrSolver(A, b; window)
     elapsed_time = ktimer(start_time)
     timemax -= elapsed_time
-    lsqr!(solver, A, b; $(kwargs_lsqr...))
+    lsqr!(solver, $(args_lsqr...); $(kwargs_lsqr...))
     solver.stats.timer += elapsed_time
     return (solver.x, solver.stats)
   end
 
-  function lsqr!(solver :: LsqrSolver{T,FC,S}, A, b :: AbstractVector{FC}; $(def_kwargs_lsqr...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}, S <: AbstractVector{FC}}
+  function lsqr!(solver :: LsqrSolver{T,FC,S}, $(def_args_lsqr...); $(def_kwargs_lsqr...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}, S <: AbstractVector{FC}}
 
     # Timer
     start_time = time_ns()

[…]
