### 2023-07-16 731c916
Use metaprogramming to reduce the code
