+++ src/cg_lanczos.jl
@@ -101,38 +101,28 @@ optargs_cg_lanczos = (:x0,)
 kwargs_cg_lanczos = (:M, :ldiv, :check_curvature, :atol, :rtol, :itmax, :timemax, :verbose, :history, :callback, :iostream)
 
 @eval begin
-  function cg_lanczos(A, b :: AbstractVector{FC}, x0 :: AbstractVector; $(def_kwargs_cg_lanczos...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}}
+  function cg_lanczos($(def_args_cg_lanczos...), $(def_optargs_cg_lanczos...); $(def_kwargs_cg_lanczos...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}}
     start_time = time_ns()
     solver = CgLanczosSolver(A, b)
-    warm_start!(solver, x0)
+    warm_start!(solver, $(optargs_cg_lanczos...))
     elapsed_time = ktimer(start_time)
     timemax -= elapsed_time
-    cg_lanczos!(solver, A, b; $(kwargs_cg_lanczos...))
+    cg_lanczos!(solver, $(args_cg_lanczos...); $(kwargs_cg_lanczos...))
     solver.stats.timer += elapsed_time
     return (solver.x, solver.stats)
   end
 
-  function cg_lanczos(A, b :: AbstractVector{FC}; $(def_kwargs_cg_lanczos...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}}
+  function cg_lanczos($(def_args_cg_lanczos...); $(def_kwargs_cg_lanczos...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}}
     start_time = time_ns()
     solver = CgLanczosSolver(A, b)
     elapsed_time = ktimer(start_time)
     timemax -= elapsed_time
-    cg_lanczos!(solver, A, b; $(kwargs_cg_lanczos...))
+    cg_lanczos!(solver, $(args_cg_lanczos...); $(kwargs_cg_lanczos...))
     solver.stats.timer += elapsed_time
     return (solver.x, solver.stats)
   end
 
-  function cg_lanczos!(solver :: CgLanczosSolver{T,FC,S}, A, b :: AbstractVector{FC}, x0 :: AbstractVector; $(def_kwargs_cg_lanczos...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}, S <: AbstractVector{FC}}
-    start_time = time_ns()
-    warm_start!(solver, x0)
-    elapsed_time = ktimer(start_time)
-    timemax -= elapsed_time
-    cg_lanczos!(solver, A, b; $(kwargs_cg_lanczos...))
-    solver.stats.timer += elapsed_time
-    return solver
-  end
-
-  function cg_lanczos!(solver :: CgLanczosSolver{T,FC,S}, A, b :: AbstractVector{FC}; $(def_kwargs_cg_lanczos...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}, S <: AbstractVector{FC}}
+  function cg_lanczos!(solver :: CgLanczosSolver{T,FC,S}, $(def_args_cg_lanczos...); $(def_kwargs_cg_lanczos...)) where {T <: AbstractFloat, FC <: FloatOrComplex{T}, S <: AbstractVector{FC}}
 
     # Timer
     start_time = time_ns()
