### 2023-07-16 f780b59
Fix tests with Apple GPUs
