### 2023-07-16 60965f3
Keep tests with AMDGPU.jl commented
