### 2023-08-03 #5187 @jedbrown
Hi @editorialbot :wave: Welcome to JOSS and thanks for agreeing to review. The instructions up top explain the review process. Your first step will be to comment `@editorialbot generate my checklist` here in the thread to create your checklist, which you can start working through. Let me know if you have any questions.
