### 2023-08-05 #5187 @LeilaGhaffari
@amontoison , this looks good to me. I have only two minor comments to add:

- [ ] Line 79: extra `s` in `theses`
- [ ] Line 80: I understand that you added this line for more clarification but it is still not quite clear in what way using BLAS affects your implementation. Is it something you can explain further?

Thanks!
