### 2023-08-05 #5187 @jedbrown
Thanks @LeilaGhaffari!

@prj- Could you please have another look to see if the remaining items on [your list](https://github.com/openjournals/joss-reviews/issues/5187#issuecomment-1445386068) are complete?
