### 2023-08-05 #5187 @prj-
All is OK, I made a follow-up comment (https://github.com/openjournals/joss-reviews/issues/5187#issuecomment-1494864299), so I'll need to double-check it has been properly addressed in the latest generated proof.
