### 2023-08-06 #5187 @dpo
@LeilaGhaffari Could you please elaborate on what is unclear? The two sentences after line 80 state

> Because the vectors manipulated by Krylov methods are always dense, the use of BLAS directly affects the efficiency of our implementations.
We also dispatch to BLAS for operator-vector products when the operator is a dense matrix.

We thought that was quite clear (BLAS1 calls for vector-vector operations and BLAS2 calls for operator-vector operations whenever the data types and storage schemes permit). We're happy to rework those sentences if you don't mind clarifying what is unclear.
