### 2023-08-06 #5187 @LeilaGhaffari
@dpo , yes, that part was clear which explains where BLAS gets used. I was mostly wondering about the quality of the effect BLAS has on your implementation. Although I got it, I was hoping to see it in clear words.

That said and reading the paragraph again, I see you started line 79 with `To speed up these procedures, ...` so I think that should clarify things. However, I find the word "affects" in `... the use of BLAS directly affects the efficiency of our implementations` a little vague/general. What do you think of using `improves` instead?
