### 2023-08-07 #5187 @amontoison
@prj- @LeilaGhaffari
I generated a new pdf.
This version should address all your comments.
