### 2023-08-08 750093c
Support StaticArrays in an extension
