### 2023-08-08 e4171ca
Actually test extensions
