+++ test/runtests.jl
@@ -45,3 +45,4 @@ include("test_mp.jl")
 include("test_solvers.jl")
 include("test_warm_start.jl")
 include("test_verbose.jl")
+include("test_extensions.jl")
