+++ Project.toml
@@ -15,8 +15,8 @@ StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 KrylovStaticArraysExt = "StaticArrays"
 
 [compat]
-julia = "^1.6.0"
 Requires = "1"
+julia = "^1.6.0"
 
 [extras]
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
