### 2023-08-08 381afcc
StaticArrays support
