+++ src/krylov_utils.jl
@@ -210,9 +210,9 @@ function ktypeof(v::S) where S <: DenseVector
 end
 
 function ktypeof(v::S) where S <: AbstractVector
-  if S.name.name == :Zeros || S.name.name == :Ones
+  if S.name.name == :Zeros || S.name.name == :Ones || S.name.name == :SArray || S.name.name == :MArray || S.name.name == :SizedArray
     T = eltype(S)
-    return Vector{T}  # FillArrays
+    return Vector{T}  # FillArrays, StaticArrays
   else
     return S  # BlockArrays, PartitionedArrays, etc...
   end
