+++ .buildkite/pipeline.yml
@@ -69,3 +69,17 @@ steps:
         Pkg.instantiate()
         include("test/gpu/metal.jl")'
     timeout_in_minutes: 30
+
+  - label: "CPUs -- StaticArrays.jl"
+    plugins:
+      - JuliaCI/julia#v1:
+          version: 1.9
+    agents:
+      queue: "juliaecosystem"
+    command: |
+      julia --color=yes --project -e '
+        using Pkg
+        Pkg.add("StaticArrays")
+        Pkg.instantiate()
+        include("test/test_extensions.jl")'
+    timeout_in_minutes: 30
