### 2023-08-08 84990dd
StaticArrays support
