+++ Project.toml
@@ -9,10 +9,12 @@ Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 
 [weakdeps]
+ComponentArrays = "b0b7db55-cfe3-40fc-9ded-d10e2dbeff66"
 FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 
 [extensions]
+KrylovComponentArraysExt = "ComponentArrays"
 KrylovFillArraysExt = "FillArrays"
 KrylovStaticArraysExt = "StaticArrays"
 
@@ -21,10 +23,11 @@ PackageExtensionCompat = "1.0.1"
 julia = "^1.6.0"
 
 [extras]
+ComponentArrays = "b0b7db55-cfe3-40fc-9ded-d10e2dbeff66"
 FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["FillArrays", "Random", "StaticArrays", "Test"]
+test = ["ComponentArrays", "FillArrays", "Random", "StaticArrays", "Test"]
