+++ ext/KrylovFillArraysExt.jl
@@ -6,7 +6,7 @@ using FillArrays: AbstractFill
 """
     Krylov.ktypeof(::AbstractFill{T,1}) where {T}
 
-Return the corresponding (non-static) `Vector{T}` type.
+Return the corresponding `Vector{T}` type.
 """
 Krylov.ktypeof(::AbstractFill{T,1}) where {T} = Vector{T}
 
