+++ ext/KrylovStaticArraysExt.jl
@@ -6,7 +6,7 @@ using StaticArrays: StaticVector
 """
     Krylov.ktypeof(::StaticVector{S,T}) where {S,T}
 
-Return the corresponding (non-static) `Vector{T}` type.
+Return the corresponding `Vector{T}` type.
 """
 Krylov.ktypeof(::StaticVector{S,T}) where {S,T} = Vector{T}
 
