### 2023-08-09 b42f714
ReshapedArrays support
