+++ src/krylov_utils.jl
@@ -7,7 +7,7 @@ const kstdout = Core.stdout
     FloatOrComplex{T}
 Union type of `T` and `Complex{T}` where T is an `AbstractFloat`.
 """
-const FloatOrComplex{T} = Union{T, Complex{T}} where T <: AbstractFloat
+const FloatOrComplex{T} = Union{T,Complex{T}} where {T<:AbstractFloat}
 
 """
     (c, s, ρ) = sym_givens(a, b)
@@ -18,40 +18,40 @@ Given `a` and `b` reals, return `(c, s, ρ)` such that
     [ c  s ] [ a ] = [ ρ ]
     [ s -c ] [ b ] = [ 0 ].
 """
-function sym_givens(a :: T, b :: T) where T <: AbstractFloat
-  #
-  # Modeled after the corresponding Matlab function by M. A. Saunders and S.-C. Choi.
-  # http://www.stanford.edu/group/SOL/dissertations/sou-cheng-choi-thesis.pdf
-  # D. Orban, Montreal, May 2015.
+function sym_givens(a::T, b::T) where {T<:AbstractFloat}
+    #
+    # Modeled after the corresponding Matlab function by M. A. Saunders and S.-C. Choi.
+    # http://www.stanford.edu/group/SOL/dissertations/sou-cheng-choi-thesis.pdf
+    # D. Orban, Montreal, May 2015.
 
-  if b == 0
-    if a == 0
-      c = one(T)
+    if b == 0
+        if a == 0
+            c = one(T)
+        else
+            c = sign(a)  # In Julia, sign(0) = 0.
+        end
+        s = zero(T)
+        ρ = abs(a)
+
+    elseif a == 0
+        c = zero(T)
+        s = sign(b)
+        ρ = abs(b)
+
+    elseif abs(b) > abs(a)
+        t = a / b
+        s = sign(b) / sqrt(one(T) + t * t)
+        c = s * t
+        ρ = b / s  # Computationally better than ρ = a / c since |c| ≤ |s|.
+
     else
-      c = sign(a)  # In Julia, sign(0) = 0.
+        t = b / a
+        c = sign(a) / sqrt(one(T) + t * t)
+        s = c * t
+        ρ = a / c  # Computationally better than ρ = b / s since |s| ≤ |c|
     end
-    s = zero(T)
-    ρ = abs(a)
 
-  elseif a == 0
-    c = zero(T)
-    s = sign(b)
-    ρ = abs(b)
-
-  elseif abs(b) > abs(a)
-    t = a / b
-    s = sign(b) / sqrt(one(T) + t * t)
-    c = s * t
-    ρ = b / s  # Computationally better than ρ = a / c since |c| ≤ |s|.
-
-  else
-    t = b / a
-    c = sign(a) / sqrt(one(T) + t * t)
-    s = c * t
-    ρ = a / c  # Computationally better than ρ = b / s since |s| ≤ |c|
-  end
-
-  return (c, s, ρ)
+    return (c, s, ρ)
 end
 
 """
@@ -62,43 +62,43 @@ c real and (s, ρ) complexes such that
     [ c   s ] [ a ] = [ ρ ]
     [ s̅  -c ] [ b ] = [ 0 ].
 """
-function sym_givens(a :: Complex{T}, b :: Complex{T}) where T <: AbstractFloat
-  #
-  # Modeled after the corresponding Fortran function by M. A. Saunders and S.-C. Choi.
-  # A. Montoison, Montreal, March 2020.
+function sym_givens(a::Complex{T}, b::Complex{T}) where {T<:AbstractFloat}
[…]
