### 2023-08-09 1d33fb6
Only for reshaped arrays that are vectors
