+++ versions.js
@@ -8,5 +8,5 @@ var DOC_VERSIONS = [
   "v0.3",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v0.9.2";
+var DOCUMENTER_NEWEST = "v0.9.3";
 var DOCUMENTER_STABLE = "stable";
