### 2023-08-12 #5187 @jedbrown
Thanks @prj-! I interpret from your statement here that you can tick the last boxes here. https://github.com/openjournals/joss-reviews/issues/5187#issuecomment-1445386068

@amontoison Thanks for your patience and this software. The revised paper looks good to me. At this point could you:
