### 2023-08-16 f6077a1
[documentation] Update tips.md
