### 2023-08-16 4b39e47
[documentation] Improve preconditioners.md
