### 2023-08-16 eb9c475
[documentation] Add an example with CuSparseMatrixCOO
