+++ test/gpu/nvidia.jl
@@ -21,9 +21,18 @@ include("gpu.jl")
     A_coo_gpu = CuSparseMatrixCOO(A_cpu)
     b_gpu = CuVector(b_cpu)
     b_gpu = CuVector(b_cpu)
-    x_csc, stats_csc = lslq(A_csc_gpu, b_gpu, verbose=1)
-    x_csr, stats_csr = lsqr(A_csr_gpu, b_gpu, verbose=1)
-    x_coo, stats_coo = lsmr(A_coo_gpu, b_gpu, verbose=1)
+    x_csc, stats_csc = lslq(A_csc_gpu, b_gpu)
+    Aᴴr_csc = A_csc_gpu' * (b_gpu - A_csc_gpu * x_csc)
+    println("cas 1")
+    println(norm(Aᴴr_csc))
+    x_csr, stats_csr = lsqr(A_csr_gpu, b_gpu)
+    Aᴴr_csr = A_csr_gpu' * (b_gpu - A_csr_gpu * x_csr)
+    println("cas 2")
+    println(norm(Aᴴr_csr))
+    x_coo, stats_coo = lsmr(A_coo_gpu, b_gpu)
+    Aᴴr_coo = A_coo_gpu' * (b_gpu - A_coo_gpu * x_coo)
+    println("cas 3")
+    println(norm(Aᴴr_coo))
 
     @testset "ic0" begin
       A_cpu, b_cpu = sparse_laplacian()
