### 2023-08-16 e25744b
Add tests for the GPU examples
