+++ test/test_verbose.jl
@@ -15,7 +15,7 @@ function test_verbose(FC)
   for fn in (:cg, :cgls, :usymqr, :cgne, :cgs, :crmr, :cg_lanczos, :dqgmres, :diom, :cr, :gpmr,
              :lslq, :lsqr, :lsmr, :lnlq, :craig, :bicgstab, :craigmr, :crls, :symmlq, :minres,
              :bilq, :minres_qlp, :qmr, :usymlq, :tricg, :trimr, :trilqr, :bilqr, :gmres, :fom,
-             :fgmres, :cg_lanczos_shift)
+             :car, :fgmres, :cg_lanczos_shift)
 
     @testset "$fn" begin
       io = IOBuffer()
