+++ src/car.jl
@@ -1,0 +1,247 @@
+# An implementation of CAR for the solution of hermitian positive definite linear systems.
+#
+# This method is described in
+#
+# A. Montoison, D. Orban and M. A. Saunders
+# MinAres: An Iterative Solver for Symmetric Linear Systems
+# Cahier du GERAD G-2023-XX.
+#
+# Alexis Montoison, <alexis.montoison@polymtl.ca>
+# Osaka, August 2023.
+
+export car, car!
+
+"""
+    (x, stats) = car(A, b::AbstractVector{FC};
+                     M=I, ldiv::Bool=false,
+                     atol::T=√eps(T), rtol::T=√eps(T),
+                     itmax::Int=0, timemax::Float64=Inf,
+                     verbose::Int=0, history::Bool=false,
+                     callback=solver->false, iostream::IO=kstdout)
+
+`T` is an `AbstractFloat` such as `Float32`, `Float64` or `BigFloat`.
+`FC` is `T` or `Complex{T}`.
+
+    (x, stats) = car(A, b, x0::AbstractVector; kwargs...)
+
+CAR can be warm-started from an initial guess `x0` where `kwargs` are the same keyword arguments as above.
+
+CAR solves the Hermitian linear system Ax = b of size n.
+
+#### Input arguments
+
+* `A`: a linear operator that models a Hermitian positive definite matrix of dimension n;
+* `b`: a vector of length n.
+
+#### Optional argument
+
+* `x0`: a vector of length n that represents an initial guess of the solution x.
+
+#### Keyword arguments
+
+* `M`: linear operator that models a Hermitian positive-definite matrix of size `n` used for centered preconditioning;
+* `ldiv`: define whether the preconditioner uses `ldiv!` or `mul!`;
+* `atol`: absolute stopping tolerance based on the residual norm;
+* `rtol`: relative stopping tolerance based on the residual norm;
+* `itmax`: the maximum number of iterations. If `itmax=0`, the default number of iterations is set to `2n`;
+* `timemax`: the time limit in seconds;
+* `verbose`: additional details can be displayed if verbose mode is enabled (verbose > 0). Information will be displayed every `verbose` iterations;
+* `history`: collect additional statistics on the run such as residual norms, or Aᴴ-residual norms;
+* `callback`: function or functor called as `callback(solver)` that returns `true` if the Krylov method should terminate, and `false` otherwise;
+* `iostream`: stream to which output is logged.
+
+#### Output arguments
+
+* `x`: a dense vector of length n;
+* `stats`: statistics collected on the run in a [`SimpleStats`](@ref) structure.
+
+#### Reference
+* A. Montoison, D. Orban and M. A. Saunders, [*MinAres: An Iterative Solver for Symmetric Linear Systems*](https://dx.doi.org/...), Cahier du GERAD G-2023-XX, GERAD, Montréal, 2023.
+"""
+function car end
+
+"""
+    solver = car!(solver::CarSolver, A, b; kwargs...)
+    solver = car!(solver::CarSolver, A, b, x0; kwargs...)
+
+where `kwargs` are keyword arguments of [`car`](@ref).
+
+See [`CarSolver`](@ref) for more details about the `solver`.
+"""
+function car! end
+
[…]
