+++ src/car.jl
@@ -4,7 +4,7 @@
 #
 # A. Montoison, D. Orban and M. A. Saunders
 # MinAres: An Iterative Solver for Symmetric Linear Systems
-# Cahier du GERAD G-2023-XX.
+# Cahier du GERAD G-2023-40.
 #
 # Alexis Montoison, <alexis.montoison@polymtl.ca>
 # Osaka, August 2023.
@@ -27,6 +27,8 @@ export car, car!
 CAR can be warm-started from an initial guess `x0` where `kwargs` are the same keyword arguments as above.
 
 CAR solves the Hermitian linear system Ax = b of size n.
+CAR minimizes ‖Arₖ‖₂ when M = Iₙ and ‖AMrₖ‖_M otherwise.
+The estimates computed every iteration are ‖Mrₖ‖₂ and ‖AMrₖ‖_M.
 
 #### Input arguments
 
@@ -56,7 +58,7 @@ CAR solves the Hermitian linear system Ax = b of size n.
 * `stats`: statistics collected on the run in a [`SimpleStats`](@ref) structure.
 
 #### Reference
-* A. Montoison, D. Orban and M. A. Saunders, [*MinAres: An Iterative Solver for Symmetric Linear Systems*](https://dx.doi.org/...), Cahier du GERAD G-2023-XX, GERAD, Montréal, 2023.
+* A. Montoison, D. Orban and M. A. Saunders, [*MinAres: An Iterative Solver for Symmetric Linear Systems*](https://dx.doi.org/...), Cahier du GERAD G-2023-40, GERAD, Montréal, 2023.
 """
 function car end
 
@@ -128,14 +130,15 @@ kwargs_car = (:M, :ldiv, :atol, :rtol, :itmax, :timemax, :verbose, :history, :ca
 
     # Tests M = Iₙ
     MisI = (M === I)
-    !MisI && error("The support of a preconditioner is not implemented in CAR.")
 
     # Check type consistency
     eltype(A) == FC || @warn "eltype(A) ≠ $FC. This could lead to errors or additional allocations in operator-vector products."
     ktypeof(b) <: S || error("ktypeof(b) is not a subtype of $S")
 
     # Set up workspace.
+    allocate_if(!MisI, solver, :Mu, S, n)
     Δx, x, r, p, s, q, t, u, stats = solver.Δx, solver.x, solver.r, solver.p, solver.s, solver.q, solver.t, solver.u, solver.stats
+    Mu = MisI ? u : solver.Mu
     warm_start = solver.warm_start
     rNorms, ArNorms = stats.residuals, stats.Aresiduals
     reset!(stats)
@@ -147,11 +150,27 @@ kwargs_car = (:M, :ldiv, :atol, :rtol, :itmax, :timemax, :verbose, :history, :ca
     else
       r .= b
     end
-    p .= r               # p₀ = r₀
-    mul!(s, A, r)        # s₀ = Ar₀
-    q .= s               # q₀ = Ap₀
-    mul!(t, A, s)        # t₀ = A²r₀
-    u .= t               # u₀ = A²p₀
+
+    # p₀ = r₀ = M(b - Ax₀)
+    if MisI
+      p .= r
+    else
+      mulorldiv!(p, M, r, ldiv)
[…]
