+++ src/minares.jl
@@ -1,0 +1,606 @@
+# An implementation of MINARES.
+#
+# This method is described in
+#
+# A. Montoison, D. Orban and M. A. Saunders
+# MinAres: An Iterative Method for Symmetric Linear Systems
+# Cahier du GERAD G-2023-40.
+#
+# Alexis Montoison, <alexis.montoison@polymtl.ca>
+# Palo Alto, March 2022.
+
+export minares, minares!
+
+"""
+    (x, stats) = minares(A, b::AbstractVector{FC};
+                         M=I, ldiv::Bool=false,
+                         λ::T = zero(T), atol::T=√eps(T),
+                         rtol::T=√eps(T), Artol::T = √eps(T),
+                         itmax::Int=0, timemax::Float64=Inf,
+                         verbose::Int=0, history::Bool=false,
+                         callback=solver->false, iostream::IO=kstdout)
+
+`T` is an `AbstractFloat` such as `Float32`, `Float64` or `BigFloat`.
+`FC` is `T` or `Complex{T}`.
+
+    (x, stats) = minares(A, b, x0::AbstractVector; kwargs...)
+
+MINARES can be warm-started from an initial guess `x0` where `kwargs` are the same keyword arguments as above.
+
+MINARES solves the Hermitian linear system Ax = b of size n.
+MINARES minimizes ‖Arₖ‖₂ when M = Iₙ and ‖AMrₖ‖_M otherwise.
+The estimates computed every iteration are ‖Mrₖ‖₂ and ‖AMrₖ‖_M.
+
+#### Input arguments
+
+* `A`: a linear operator that models a Hermitian positive definite matrix of dimension n;
+* `b`: a vector of length n.
+
+#### Optional argument
+
+* `x0`: a vector of length n that represents an initial guess of the solution x.
+
+#### Keyword arguments
+
+* `M`: linear operator that models a Hermitian positive-definite matrix of size `n` used for centered preconditioning;
+* `ldiv`: define whether the preconditioner uses `ldiv!` or `mul!`;
+* `λ`: regularization parameter;
+* `atol`: absolute stopping tolerance based on the residual norm;
+* `rtol`: relative stopping tolerance based on the residual norm;
+* `Artol`: relative stopping tolerance based on the Aᴴ-residual norm;
+* `itmax`: the maximum number of iterations. If `itmax=0`, the default number of iterations is set to `2n`;
+* `timemax`: the time limit in seconds;
[…]
