+++ src/Krylov.jl
@@ -16,6 +16,7 @@ include("cg_lanczos.jl")
 include("cg_lanczos_shift.jl")
 include("minres.jl")
 include("minres_qlp.jl")
+include("minares.jl")
 
 include("diom.jl")
 include("fom.jl")
