### 2023-09-01 0989990
Add an example for MINARES
