### 2023-09-01 6cee319
Add a CITATION.cff
