+++ CITATION.cff
@@ -1,0 +1,38 @@
+cff-version: 1.2.0
+title: >-
+  Krylov.jl: A Julia Basket of Hand-Picked Krylov Methods
+message: >-
+  If you use this software, please cite it using the
+  metadata from this file.
+type: software
+authors:
+  - given-names: Alexis
+    family-names: Montoison
+    email: alexis.montoison@polymtl.ca
+    affiliation: >-
+      GERAD and Department of Mathematics and
+      Industrial Engineering, Polytechnique Montréal,
+      QC, Canada.
+    orcid: 'https://orcid.org/0000-0002-3403-5450'
+  - given-names: Dominique
+    family-names: Orban
+    email: dominique.orban@gerad.ca
+    orcid: 'https://orcid.org/0000-0002-8017-7687'
+    affiliation: >-
+      GERAD and Department of Mathematics and
+      Industrial Engineering, Polytechnique Montréal,
+      QC, Canada
+identifiers:
+  - description: Zenodo archive
+    type: doi
+    value: 10.5281/zenodo.822073
+keywords:
+  - Julia
+  - linear algebra
+  - Krylov methods
+  - Krylov processes
+  - sparse linear systems
+  - GPU computing
+license: MPL-2.0
+repository-code: >-
+  https://github.com/JuliaSmoothOptimizers/Krylov.jl
