+++ CITATION.bib
@@ -1,8 +1,0 @@
-@Misc{montoison-orban-krylov-2020,
-  author = {A. Montoison and D. Orban and {contributors}},
-  title = {{Krylov.jl}: A {J}ulia Basket of Hand-Picked {K}rylov Methods},
-  month = {June},
-  howpublished = {https://github.com/JuliaSmoothOptimizers/Krylov.jl},
-  year = {2020},
-  DOI = {10.5281/zenodo.822073},
-}
