### 2023-09-01 #5187 @amontoison
@jedbrown

>     * [ ]  Make a tagged release of your software, and list the version tag of the archived version here.
The version tag is [`v0.9.4`](https://github.com/JuliaSmoothOptimizers/Krylov.jl/tree/v0.9.4).
https://github.com/JuliaSmoothOptimizers/Krylov.jl/releases/tag/v0.9.4

>     * [ ]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)

I archived the reviewed software in Zenodo: https://zenodo.org/record/8310030

>     * [ ]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.

I checked the metadata on Zenodo. :white_check_mark:

>     * [ ]  Please list the DOI of the archived version here.

The DOI of the release 0.9.4 is `10.5281/zenodo.8310030`.
