### 2023-09-07 04757ec
[documentation] AMDGPU.jl supports rocSPARSE
