### 2023-09-07 7ff7abb
[documentation] AMDGPU.jl supports rocSPARSE
