### 2023-09-07 c64dc49
Fix a typo in test/gpu/amd.jl
