+++ test/gpu/amd.jl
@@ -22,7 +22,6 @@ include("gpu.jl")
     A_csc_gpu = ROCSparseMatrixCSC(A_cpu)
     A_csr_gpu = ROCSparseMatrixCSR(A_cpu)
     A_coo_gpu = ROCSparseMatrixCOO(A_cpu)
-    b_gpu = ROXVector(b_cpu)
     b_gpu = ROCVector(b_cpu)
     x_csc, y_csc, stats_csc = lnlq(A_csc_gpu, b_gpu)
     x_csr, y_csr, stats_csr = craig(A_csr_gpu, b_gpu)
