### 2023-09-09 d0b0ae0
[documentation] Update the section about AMD GPUs
