### 2023-09-11 82a7eb1
Improve the function hermitian_lanczos
