### 2023-09-17 d956bcd
[documentation] Explain the transpose-free variant of the non-Hermitian Lanczos process
