### 2023-09-17 7f39e2d
[documentation] Explain the transpose-free variant of the non-Hermitian Lanczos process
