+++ dev/assets/documenter.js
@@ -1,15 +1,15 @@
 // Generated by Documenter.jl
 requirejs.config({
   paths: {
-    'highlight-julia': 'https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.5.1/languages/julia.min',
+    'highlight-julia': 'https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.0.1/languages/julia.min',
     'headroom': 'https://cdnjs.cloudflare.com/ajax/libs/headroom/0.12.0/headroom.min',
     'jqueryui': 'https://cdnjs.cloudflare.com/ajax/libs/jqueryui/1.12.1/jquery-ui.min',
-    'katex-auto-render': 'https://cdnjs.cloudflare.com/ajax/libs/KaTeX/0.13.24/contrib/auto-render.min',
+    'katex-auto-render': 'https://cdnjs.cloudflare.com/ajax/libs/KaTeX/0.13.11/contrib/auto-render.min',
     'jquery': 'https://cdnjs.cloudflare.com/ajax/libs/jquery/3.6.0/jquery.min',
     'headroom-jquery': 'https://cdnjs.cloudflare.com/ajax/libs/headroom/0.12.0/jQuery.headroom.min',
-    'katex': 'https://cdnjs.cloudflare.com/ajax/libs/KaTeX/0.13.24/katex.min',
-    'highlight': 'https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.5.1/highlight.min',
-    'highlight-julia-repl': 'https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.5.1/languages/julia-repl.min',
+    'katex': 'https://cdnjs.cloudflare.com/ajax/libs/KaTeX/0.13.11/katex.min',
+    'highlight': 'https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.0.1/highlight.min',
+    'highlight-julia-repl': 'https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.0.1/languages/julia-repl.min',
   },
   shim: {
   "highlight-julia": {
@@ -69,67 +69,6 @@ require(['jquery', 'highlight', 'highlight-julia', 'highlight-julia-repl'], func
 $(document).ready(function() {
     hljs.highlightAll();
 })
-
-})
-////////////////////////////////////////////////////////////////////////////////
-require([], function() {
-function addCopyButtonCallbacks() {
-  for (const el of document.getElementsByTagName("pre")) {
-    const button = document.createElement("button");
-    button.classList.add("copy-button", "fas", "fa-copy");
-    el.appendChild(button);
-
-    const success = function () {
-      button.classList.add("success", "fa-check");
-      button.classList.remove("fa-copy");
-    };
-
-    const failure = function () {
-      button.classList.add("error", "fa-times");
-      button.classList.remove("fa-copy");
-    };
-
-    button.addEventListener("click", function () {
-      copyToClipboard(el.innerText).then(success, failure);
-
-      setTimeout(function () {
-        button.classList.add("fa-copy");
-        button.classList.remove("success", "fa-check", "fa-times");
-      }, 5000);
-    });
-  }
-}
-
-function copyToClipboard(text) {
-  // clipboard API is only available in secure contexts
-  if (window.navigator && window.navigator.clipboard) {
-    return window.navigator.clipboard.writeText(text);
-  } else {
-    return new Promise(function (resolve, reject) {
-      try {
-        const el = document.createElement("textarea");
-        el.textContent = text;
-        el.style.position = "fixed";
-        el.style.opacity = 0;
-        document.body.appendChild(el);
-        el.select();
-        document.execCommand("copy");
-
-        resolve();
-      } catch (err) {
-        reject(err);
[…]
