### 2023-09-21 #5187 @amontoison
I found a small bug with the Julia environment, after a `'` (adjoint), the color of the code is modified.
![Capture d’écran du 2023-09-21 00-36-53](https://github.com/openjournals/joss-reviews/assets/35051714/e7911dbb-3a12-4086-ad55-0e1d1f69ddba)
