### 2023-09-21 #5187 @danielskatz
@amontoison - I'm the track editor who will handle the rest of the process. Later today, I'll proofread this and let you know next steps.

In the meantime, I don't understand the impact of your comment. Are you saying that you need to make a change to the .md file to fix this?  If so, please go ahead and then use `@editorialbot generate pdf` as a new comment to generate a new PDF and check it.  If not, please let me know what action you need.
