### 2023-09-21 #5187 @amontoison
@danielskatz - Sorry if my comment was not clear. I just remarked that a Julia block of code (```'''julia ... '''```) is not well displayed.
```julia
# Solve Py = x
function ldiv_ic0!(P, x, y, z)
  ldiv!(z, LowerTriangular(P), x)   # Forward substitution with L
  ldiv!(y, LowerTriangular(P)', z)  # Backward substitution with Lᴴ
  return y
end
```
The issue is related to the lstlisting used by JOSS for the Julia language and not my Markdown file.
