### 2023-09-22 #5187 @danielskatz
Do you want us to work on this internally, or should we go ahead and publish, knowing this issue exists?
