### 2023-09-22 #5187 @danielskatz
Also, I've proofread the paper, and have a few suggestions, as shown in https://github.com/JuliaSmoothOptimizers/Krylov.jl/pull/809 - please merge this, or let me know what you disagree with.
