### 2023-09-22 acefe34
Fix the issue with the adjoint in JOSS
