### 2023-09-23 #5187 @amontoison
> Do you want us to work on this internally, or should we go ahead and publish, knowing this issue exists?

I reformulated a little bit the examples to avoid this issue (https://github.com/JuliaSmoothOptimizers/Krylov.jl/pull/810).

I will let @dpo review https://github.com/JuliaSmoothOptimizers/Krylov.jl/pull/809 for the final tweaks.
