+++ previews/PR800/assets/documenter.js
@@ -1,15 +1,16 @@
 // Generated by Documenter.jl
 requirejs.config({
   paths: {
-    'highlight-julia': 'https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.0.1/languages/julia.min',
+    'highlight-julia': 'https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.8.0/languages/julia.min',
     'headroom': 'https://cdnjs.cloudflare.com/ajax/libs/headroom/0.12.0/headroom.min',
-    'jqueryui': 'https://cdnjs.cloudflare.com/ajax/libs/jqueryui/1.12.1/jquery-ui.min',
-    'katex-auto-render': 'https://cdnjs.cloudflare.com/ajax/libs/KaTeX/0.13.11/contrib/auto-render.min',
-    'jquery': 'https://cdnjs.cloudflare.com/ajax/libs/jquery/3.6.0/jquery.min',
+    'jqueryui': 'https://cdnjs.cloudflare.com/ajax/libs/jqueryui/1.13.2/jquery-ui.min',
+    'minisearch': 'https://cdn.jsdelivr.net/npm/minisearch@6.1.0/dist/umd/index.min',
+    'katex-auto-render': 'https://cdnjs.cloudflare.com/ajax/libs/KaTeX/0.16.8/contrib/auto-render.min',
+    'jquery': 'https://cdnjs.cloudflare.com/ajax/libs/jquery/3.7.0/jquery.min',
     'headroom-jquery': 'https://cdnjs.cloudflare.com/ajax/libs/headroom/0.12.0/jQuery.headroom.min',
-    'katex': 'https://cdnjs.cloudflare.com/ajax/libs/KaTeX/0.13.11/katex.min',
-    'highlight': 'https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.0.1/highlight.min',
-    'highlight-julia-repl': 'https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.0.1/languages/julia-repl.min',
+    'katex': 'https://cdnjs.cloudflare.com/ajax/libs/KaTeX/0.16.8/katex.min',
+    'highlight': 'https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.8.0/highlight.min',
+    'highlight-julia-repl': 'https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.8.0/languages/julia-repl.min',
   },
   shim: {
   "highlight-julia": {
@@ -72,180 +73,789 @@ $(document).ready(function() {
 
 })
 ////////////////////////////////////////////////////////////////////////////////
+require(['jquery'], function($) {
+
+var isExpanded = true;
+
+$(document).on("click", ".docstring header", function () {
+  let articleToggleTitle = "Expand docstring";
+
+  if ($(this).siblings("section").is(":visible")) {
+    $(this)
+      .find(".docstring-article-toggle-button")
+      .removeClass("fa-chevron-down")
+      .addClass("fa-chevron-right");
+  } else {
+    $(this)
+      .find(".docstring-article-toggle-button")
+      .removeClass("fa-chevron-right")
+      .addClass("fa-chevron-down");
+
+    articleToggleTitle = "Collapse docstring";
+  }
+
+  $(this)
+    .find(".docstring-article-toggle-button")
+    .prop("title", articleToggleTitle);
+  $(this).siblings("section").slideToggle();
+});
+
+$(document).on("click", ".docs-article-toggle-button", function () {
+  let articleToggleTitle = "Expand docstring";
+  let navArticleToggleTitle = "Expand all docstrings";
+
+  if (isExpanded) {
+    $(this).removeClass("fa-chevron-up").addClass("fa-chevron-down");
[…]
