+++ src/krylov_processes.jl
@@ -1,7 +1,7 @@
 export hermitian_lanczos, nonhermitian_lanczos, arnoldi, golub_kahan, saunders_simon_yip, montoison_orban
 
 """
-    V, T = hermitian_lanczos(A, b, k)
+    V, β, T = hermitian_lanczos(A, b, k)
 
 #### Input arguments
 
@@ -12,6 +12,7 @@ export hermitian_lanczos, nonhermitian_lanczos, arnoldi, golub_kahan, saunders_s
 #### Output arguments
 
 * `V`: a dense n × (k+1) matrix;
+* `β`: a coefficient such that βv₁ = b;
 * `T`: a sparse (k+1) × k tridiagonal matrix.
 
 #### Reference
@@ -41,6 +42,7 @@ function hermitian_lanczos(A, b::AbstractVector{FC}, k::Int) where FC <: FloatOr
     end
   end
 
+  local β₁
   V = M(undef, n, k+1)
   T = SparseMatrixCSC(k+1, k, colptr, rowval, nzval)
 
@@ -49,8 +51,8 @@ function hermitian_lanczos(A, b::AbstractVector{FC}, k::Int) where FC <: FloatOr
     vᵢ = view(V,:,i)
     vᵢ₊₁ = q = view(V,:,i+1)
     if i == 1
-      βᵢ = @knrm2(n, b)
-      vᵢ .= b ./ βᵢ
+      β₁ = @knrm2(n, b)
+      vᵢ .= b ./ β₁
     end
     mul!(q, A, vᵢ)
     if i ≥ 2
@@ -67,11 +69,11 @@ function hermitian_lanczos(A, b::AbstractVector{FC}, k::Int) where FC <: FloatOr
     vᵢ₊₁ .= q ./ βᵢ₊₁
     pαᵢ = pαᵢ + 3
   end
-  return V, T
+  return V, β₁, T
 end
 
 """
-    V, T, U, Tᴴ = nonhermitian_lanczos(A, b, c, k)
+    V, β, T, U, γᴴ, Tᴴ = nonhermitian_lanczos(A, b, c, k)
 
 #### Input arguments
 
@@ -83,8 +85,10 @@ end
 #### Output arguments
 
 * `V`: a dense n × (k+1) matrix;
+* `β`: a coefficient such that βv₁ = b;
 * `T`: a sparse (k+1) × k tridiagonal matrix;
 * `U`: a dense n × (k+1) matrix;
+* `γᴴ`: a coefficient such that γᴴu₁ = c;
 * `Tᴴ`: a sparse (k+1) × k tridiagonal matrix.
 
 #### Reference
@@ -115,6 +119,7 @@ function nonhermitian_lanczos(A, b::AbstractVector{FC}, c::AbstractVector{FC}, k
     end
   end
 
+  local β₁, γ₁ᴴ
   V = M(undef, n, k+1)
   U = M(undef, n, k+1)
   T = SparseMatrixCSC(k+1, k, colptr, rowval, nzval_T)
@@ -128,10 +133,10 @@ function nonhermitian_lanczos(A, b::AbstractVector{FC}, c::AbstractVector{FC}, k
     uᵢ₊₁ = p = view(U,:,i+1)
     if i == 1
       cᴴb = @kdot(n, c, b)
-      βᵢ = √(abs(cᴴb))
-      γᵢ = cᴴb / βᵢ
-      vᵢ .= b ./ βᵢ
-      uᵢ .= c ./ conj(γᵢ)
+      β₁ = √(abs(cᴴb))
+      γ₁ᴴ = conj(cᴴb / β₁)
+      vᵢ .= b ./ β₁
+      uᵢ .= c ./ γ₁ᴴ
     end
[…]
