### 2023-09-24 a06de4b
[Tests] Update gpu.jl
